A technical working prototype creating namelists from a yml template, check consistency and input data based a generic template populated from namelist values
