+++ examples/binary_input_data.json
@@ -1,0 +1,122 @@
+{
+    "pgd": {
+        "NAM_FRAC#LECOSG": {
+            "True": {
+                "macros": {
+                    "VTYPE": {
+                        "fmt": "02d"
+                    },
+                    "DECADE": {
+                        "ntime": "NAM_DATA_ISBA#NTIME"
+                    }
+                },
+                "NAM_DATA_ISBA#CFTYP_ALBNIR_SOIL": {
+                    "DIRTYPE": {
+                        "NAM_DATA_ISBA#CFNAM_ALBNIR_SOIL": "@ecoclimap_sg@/ALB_@VTYPE@_@DECADE@"
+                    }
+                }
+            },
+            "False": {
+                "ecoclimapI_covers_param.bin": "@ecoclimap_bin_dir@/ecoclimapI_covers_param.bin",
+                "ecoclimapII_eu_covers_param.bin": "@ecoclimap_bin_dir@/ecoclimapII_eu_covers_param.bin",
+                "ecoclimapII_af_covers_param.bin": "@ecoclimap_bin_dir@/ecoclimapII_af_covers_param.bin"
+            }
+        },
+        "NAM_ZS#YFILETYPE": {
+            "DIRECT": {
+                "NAM_ZS#YFILE": "@gmted@/gmted2010"
+            }
+        }
+    },
+    "prep":{
+
+    },
+    "perturbed":{
+        "NAM_FRAC#LECOSG": {
+            "False": {
+                "ecoclimapI_covers_param.bin": "@ecoclimap_bin_dir@/ecoclimapI_covers_param.bin",
+                "ecoclimapII_eu_covers_param.bin": "@ecoclimap_bin_dir@/ecoclimapII_eu_covers_param.bin",
+                "ecoclimapII_af_covers_param.bin": "@ecoclimap_bin_dir@/ecoclimapII_af_covers_param.bin"
+            }
+        }
+    },
+    "offline":{
+        "NAM_FRAC#LECOSG": {
+            "False": {
+                "ecoclimapI_covers_param.bin": "@ecoclimap_bin_dir@/ecoclimapI_covers_param.bin",
+                "ecoclimapII_eu_covers_param.bin": "@ecoclimap_bin_dir@/ecoclimapII_eu_covers_param.bin",
+                "ecoclimapII_af_covers_param.bin": "@ecoclimap_bin_dir@/ecoclimapII_af_covers_param.bin"
+            }
+        }
+    },
+    "soda": {
+        "NAM_ASSIM#CASSIM_ISBA": {
+            "OI": {
+                "NAM_ASSIM#CFILE_FORMAT_FG": {
+                    "FA": {
+                        "FG_OI_MAIN": "@first_guess_dir@/FG_OI_MAIN"
+                    },
+                    "NC": {
+                        "FIRST_GUESS_@YY@@MM@@DD@H@HH@.DAT": ""
+                    }
+                },
+                "ASCAT_SM.DAT": "@ascat_dir@/ASCAT_SM.DAT",
+                "fort.61": "@oi_coeffs_dir@/ISBA_POLYNOMES",
+                "NAM_ASSIM#CFILE_FORMAT_LSM": {
+                    "FA": {
+                        "FG_OI_MAIN": "@first_guess_dir@/FG_OI_MAIN"
+                    },
+                    "ASCII": {
+                        "LSM.DAT": "@climdir@/LSM.DAT"
+                    }
+                }
+            },
+            "EKF": {
+                "extenders": {
+                    "PERT": {
+                        "ekfpert": {
+                            "list": "NAM_VAR#NNCV",
+                            "duplicate": "NAM_ASSIM#LLINCHECK"
+                        }
+                    }
+                },
+                "NAM_IO_OFFLINE#CSURFFILETYPE": {
+                    "FA": {
+                        "PREP_INIT.fa": "@first_guess_dir@/PREP_INIT.fa",
+                        "PREP_@FG_YY@@FG_MM@@FG_DD@H@FG_HH@_EKF_PERT@PERT@.fa": "@first_guess_dir@/PREP_@FG_YY@@FG_MM@@FG_DD@H@FG_HH@_EKF_PERT@PERT@.fa"
+                    },
+                    "NC": {
+                        "PREP_INIT.nc": "@first_guess_dir@/PREP_INIT.nc",
+                        "PREP_@FG_YY@@FG_MM@@FG_DD@H@FG_HH@_EKF_PERT@PERT@.nc": "@first_guess_dir@/PREP_@FG_YY@@FG_MM@@FG_DD@H@FG_HH@_EKF_PERT@PERT@.nc"
+                    }
+                } 
+            },
+            "ENKF": {
+                "extenders": {
+                    "EE": {
+                        "iterator": {
+                            "start": 0,
+                            "end": "NAM_ASSIM#NENS_M",
+                            "step": 1,
+                            "fmt": "{:02d}"
+                        }
+                    }
+                },
+                "NAM_IO_OFFLINE#CSURFFILETYPE": {
+                    "FA": {
+                        "PREP_INIT.fa": "@first_guess_dir@/PREP_INIT.fa",
+                        "PREP_@FG_YY@@FG_MM@@FG_DD@H@FG_HH@_EKF_ENS@EE@.fa": "@first_guess_dir@/PREP_@FG_YY@@FG_MM@@FG_DD@H@FG_HH@_EKF_ENS@EE@.fa"
+                    }
+                },
+                "NAM_ASSIM#CFILE_FORMAT_LSM": {
+                    "FA": {
+                        "FG_OI_MAIN": "@first_guess_dir@/FG_OI_MAIN"
+                    },
+                    "ASCII": {
+                        "LSM.DAT": "@first_guess_dir@/LSM.DAT"
+                    }
+                }
+            }
+        }        
+    }
+}
