+++ pysurfex/namelist.py
@@ -1,1227 +1,241 @@
 """Namelist."""
-import json
 import logging
-import os
+import re
 
 import f90nml
-import yaml
 
-from .datetime_utils import as_datetime, as_timedelta
-from .ecoclimap import Ecoclimap, EcoclimapSG
+from .binary_input import InputDataFromNamelist
 
 
-class BaseNamelist(object):
-    """Base namelist."""
+class NamelistGenerator(object):
+    """Namelist class."""
 
-    def __init__(
-        self,
-        program,
-        config,
-        input_path,
-        forc_zs=False,
-        prep_file=None,
-        prep_filetype=None,
-        prep_pgdfile=None,
-        prep_pgdfiletype=None,
-        dtg=None,
-        fcint=3,
-        geo=None,
-    ):
-        """Construct a base namelists class to be implemented by namelist implementations.
+    def __init__(self, program, config, definitions, assemble=None, consistency=True):
+        """Construct a base namelists class.
 
         Args:
             program (str): Which surfex binary you want to run ["pgd", "prep", "offline", "soda"]
             config (surfex.Configuration): A SURFEX configuration object
-            input_path (str): A path to search for json namelist blocks
-            forc_zs (bool): Set the surfex orography heigth to the same as in the forcing files
-            prep_file (str): Input file for prep
-            prep_filetype (str): The format of prep_file
-            prep_pgdfile (str): Input PGD file for prep in case it comes from SURFEX
-            prep_pgdfiletype (str): The format of the prep_pgdfile
-            dtg (datetime.datetime): The date/time you want to run
-            fcint (int): The intervall between the cycles. Used for first guesses.
-            geo (surfex.Geo): Surfex geometry. The domain you want to run on
-
-        Raises:
-            RuntimeError: Needed input
-            NotImplementedError: Mode not implemented
-
+            definitions (dict): Namelist definitions
+            assemble(dict, optional): Assembly order. Default to None.
+            consistency (bool, optional): Check configuration consistency. Defaults to True.
         """
+        self.program = program
         self.config = config
-        self.input_path = input_path
-        self.forc_zs = forc_zs
-        if dtg is not None:
-            if isinstance(dtg, str):
-                dtg = as_datetime(dtg)
-        self.dtg = dtg
-        check_parsing = True
-        if self.dtg is None:
-            check_parsing = False
-        self.fcint = fcint
-        self.geo = geo
-
-        # The time stamp of next cycle file
-        forecast_length = self.fcint
-        if self.dtg is not None:
-            self.end_of_forecast = self.dtg + as_timedelta(seconds=forecast_length * 3600)
-        else:
-            self.end_of_forecast = None
-
-        logging.info("Creating JSON namelist input for program: %s", program)
-        self.input_list = []
-
-        self.prolog(check_parsing)
-        # Program specific settings
-        if program == "pgd":
-            self.set_pgd_namelist()
-        elif program == "prep":
-            if prep_file is None:
-                raise RuntimeError(
-                    "Prep need an input file either as a json namelist or a surfex "
-                    "supported format"
-                )
-            self.set_prep_namelist(
-                prep_file=prep_file,
-                prep_filetype=prep_filetype,
-                prep_pgdfile=prep_pgdfile,
-                prep_pgdfiletype=prep_pgdfiletype,
-            )
-        elif program == "offline" or program == "perturbed":
-            self.set_offline_namelist()
-        elif program == "soda":
-            self.set_soda_namelist()
-        else:
-            raise NotImplementedError(program)
-        self.epilog()
-        self.override()
-
-    def prolog(self, check_parsing):
-        """Prolog.
-
-        Args:
-            check_parsing (_type_): _description_
-
-        """
-        # IO
-        self.input_list.append({"file": self.input_path + "/io.json"})
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_IO_OFFLINE": {
-                        "CSURF_FILETYPE": self.config.get_setting(
-                            "SURFEX#IO#CSURF_FILETYPE"
-                        )
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_IO_OFFLINE": {
-                        "CTIMESERIES_FILETYPE": self.config.get_setting(
-                            "SURFEX#IO#CTIMESERIES_FILETYPE"
-                        )
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_IO_OFFLINE": {
-                        "CFORCING_FILETYPE": self.config.get_setting(
-                            "SURFEX#IO#CFORCING_FILETYPE"
-                        )
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_IO_OFFLINE": {
-                        "CPGDFILE": self.config.get_setting("SURFEX#IO#CPGDFILE")
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_IO_OFFLINE": {
-                        "CPREPFILE": self.config.get_setting("SURFEX#IO#CPREPFILE")
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_IO_OFFLINE": {
-                        "CSURFFILE": self.config.get_setting(
-                            "SURFEX#IO#CSURFFILE",
-                            validtime=self.end_of_forecast,
-                            basedtg=self.dtg,
-                            check_parsing=check_parsing,
-                        )
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_IO_OFFLINE": {
-                        "XTSTEP_SURF": self.config.get_setting("SURFEX#IO#XTSTEP")
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_IO_OFFLINE": {
-                        "XTSTEP_OUTPUT": self.config.get_setting(
-                            "SURFEX#IO#XTSTEP_OUTPUT"
-                        )
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_WRITE_SURF_ATM": {
-                        "LSPLIT_PATCH": self.config.get_setting("SURFEX#IO#LSPLIT_PATCH")
-                    }
-                }
-            }
-        )
-
-        if self.forc_zs:
-            self.input_list.append({"json": {"NAM_IO_OFFLINE": {"LSET_FORC_ZS": True}}})
-
-        # Constants and parameters
-        self.input_list.append({"file": self.input_path + "/constants.json"})
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_SURF_ATM": {
-                        "XRIMAX": self.config.get_setting("SURFEX#PARAMETERS#XRIMAX")
-                    }
-                }
-            }
-        )
-
-    def set_pgd_namelist(self):
-        """Set pgd namelist."""
-        # PGS schemes
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_PGD_SCHEMES": {
-                        "CSEA": self.config.get_setting("SURFEX#TILES#SEA"),
-                        "CWATER": self.config.get_setting("SURFEX#TILES#INLAND_WATER"),
-                        "CNATURE": self.config.get_setting("SURFEX#TILES#NATURE"),
-                        "CTOWN": self.config.get_setting("SURFEX#TILES#TOWN"),
-                    }
-                }
-            }
-        )
-
-        eco_sg = self.config.get_setting("SURFEX#COVER#SG")
-        # Ecoclimap SG
-        self.input_list.append({"json": {"NAM_FRAC": {"LECOSG": eco_sg}}})
-        if self.config.get_setting("SURFEX#COVER#SG"):
-            ecoclimap = EcoclimapSG(self.config)
-
-            self.input_list.append(
-                {"json": {"NAM_DATA_ISBA": {"NTIME": ecoclimap.decades}}}
-            )
-
-            fname = self.config.get_setting("SURFEX#COVER#H_TREE")
-            if fname != "" and fname is not None:
-                self.input_list.append(
-                    self.set_dirtyp_data_namelist(
-                        "NAM_DATA_ISBA", "H_TREE", fname, vtype=1
-                    )
-                )
-
-            decadal_data_types = [
-                "ALBNIR_SOIL",
-                "ALBNIR_VEG",
-                "ALBVIS_SOIL",
-                "ALBVIS_VEG",
-                "LAI",
-            ]
-            for decadal_data_type in decadal_data_types:
-                for vtt in range(1, ecoclimap.veg_types + 1):
-                    for decade in range(1, ecoclimap.decades + 1):
-                        filepattern = self.config.get_setting(
-                            "SURFEX#COVER#" + decadal_data_type, check_parsing=False
-                        )
-                        fname = ecoclimap.parse_fnames(filepattern, decade)
-                        self.input_list.append(
-                            self.set_dirtyp_data_namelist(
-                                "NAM_DATA_ISBA",
-                                decadal_data_type,
-                                fname,
-                                vtype=vtt,
-                                decade=decade,
-                            )
-                        )
-
-        ecoclimap_dir = "ecoclimap_dir"
-        if self.config.get_setting("SURFEX#COVER#SG"):
-            ecoclimap_dir = "ecoclimap_sg_cover_dir"
-
-        possible_direct_data = {
-            "ISBA": {
-                "YSAND": "sand_dir",
-                "YCLAY": "clay_dir",
-                "YSOC_TOP": "soc_top_dir",
-                "YSOC_SUB": "soc_sub_dir",
-            },
-            "COVER": {"YCOVER": ecoclimap_dir},
-            "ZS": {"YZS": "oro_dir"},
+        self.nldict = definitions
+        macros_defs = {
+            "CPGDFILE": "SURFEX#IO#CPGDFILE",
+            "CPREPFILE": "SURFEX#IO#CPREPFILE",
+            "CSURFFILE": "SURFEX#IO#CSURFFILE",
+            "CSURF_FILETYPE": "SURFEX#IO#CSURF_FILETYPE",
+            "CTIMESERIES_FILETYPE": "SURFEX#IO#CTIMESERIES_FILETYPE",
+            "XRIMAX": "SURFEX#PARAMETERS#XRIMAX",
+            "LSPLIT_PATCH": "SURFEX#IO#LSPLIT_PATCH"
         }
-        for namelist_section, ftypes in possible_direct_data.items():
-            # for ftype in possible_direct_data[namelist_section]:
-            for ftype in ftypes:
-                fname = str(
-                    self.config.get_setting("SURFEX#" + namelist_section + "#" + ftype)
-                )
-                self.input_list.append(
-                    self.set_direct_data_namelist(
-                        "NAM_" + namelist_section, ftype, fname, self.input_path
-                    )
-                )
+        macros = {}
+        for macro, setting in macros_defs.items():
+            vmacro = config.get_setting(setting)
+            logging.debug("Mapping macro %s = %s", macro, vmacro)
+            if vmacro is not None:
+                macros.update({
+                    macro: vmacro
+                })
+        self.macros = macros
+        if assemble is None:
+            self.assemble = self.namelist_blocks()
+        else:
+            self.assemble = assemble
+        logging.debug(self.assemble)
+        nlres = self.assemble_namelist()
+        self.nml = f90nml.Namelist(nlres)
+        if consistency:
+            problems = self.concistency(self.nml)
+            if problems:
+                logging.warning("Found problems!")
+                for problem, info in problems.items():
+                    for level, desc in info.items():
+                        if level == "SEVERE":
+                            logging.error(
+                                "Problem with: %s Description=%s", problem, desc
+                            )
+                        elif level == "WARNING":
+                            logging.warning(
+                                "Problem with: %s Description=%s", problem, desc
+                            )
+                        else:
+                            logging.info(
+                                "Problem with: %s Description=%s", problem, desc
+                            )
 
-        # Set ISBA properties
+    def namelist_blocks(self):
+        """Construct building blocks for the namelist genrator."""
+        logging.info("Building namelist blocks for program: %s", self.program)
+
+        input_blocks = ["io", "constants", "sice", "treedrag", "flake"]
+        lisba = False
         if self.config.get_setting("SURFEX#ISBA#SCHEME") == "DIF":
-            self.input_list.append(
-                {"json": {"NAM_ISBA": {"CISBA": "DIF", "NGROUND_LAYER": 14}}}
-            )
-            if os.path.exists(self.input_path + "/isba_dif.json"):
-                self.input_list.append({"file": self.input_path + "/isba_dif.json"})
+            lisba = True
         elif self.config.get_setting("SURFEX#ISBA#SCHEME") == "3-L":
-            self.input_list.append(
-                {"json": {"NAM_ISBA": {"CISBA": "3-L", "NGROUND_LAYER": 3}}}
-            )
+            lisba = True
         elif self.config.get_setting("SURFEX#ISBA#SCHEME") == "2-L":
-            self.input_list.append(
-                {"json": {"NAM_ISBA": {"CISBA": "2-L", "NGROUND_LAYER": 2}}}
-            )
+            lisba = True
 
-        # Set patches
-        npatch = self.config.get_setting("SURFEX#ISBA#NPATCH")
-        self.input_list.append({"json": {"NAM_ISBA": {"NPATCH": npatch}}})
-
-        # Set MEB
-        lmeb = self.config.get_setting("SURFEX#ISBA#MEB")
-        self.input_list.append({"json": {"NAM_ISBA": {"LMEB": lmeb}}})
-        if lmeb:
-            self.input_list.append({"file": self.input_path + "/meb_settings.json"})
-
-        # RSMIN
-        if self.config.get_setting("SURFEX#COVER#SG"):
-            self.input_list.append({"file": self.input_path + "/rsmin_sg.json"})
-            self.input_list.append({"file": self.input_path + "/rsmin_sg_mod.json"})
-        else:
-            self.input_list.append({"file": self.input_path + "/rsmin.json"})
-            self.input_list.append({"file": self.input_path + "/rsmin_mod.json"})
-
-        # CV
-        if self.config.get_setting("SURFEX#COVER#SG"):
-            self.input_list.append({"file": self.input_path + "/cv_sg.json"})
-        else:
-            self.input_list.append({"file": self.input_path + "/cv.json"})
-
-        # Treedrag
-        if self.config.get_setting("SURFEX#TREEDRAG#TREEDATA_FILE") != "":
-            treeheight = self.config.get_setting("SURFEX#TREEDRAG#TREEDATA_FILE")
-            self.input_list.append(
-                {
-                    "json": {
-                        "NAM_DATA_ISBA": {
-                            "CFNAM_H_TREE(4)": treeheight,
-                            "CFTYP_H_TREE(4)": "ASCLLV",
-                            "CFNAM_H_TREE(5)": treeheight,
-                            "CFTYP_H_TREE(5)": "ASCLLV",
-                            "CFNAM_H_TREE(6)": treeheight,
-                            "CFTYP_H_TREE(6)": "ASCLLV",
-                        }
-                    }
-                }
-            )
-
-        if self.config.get_setting("SURFEX#TOWN#LTOWN_TO_ROCK"):
-            if self.config.get_setting("SURFEX#TILES#TOWN") != "NONE":
-                logging.warning(
-                    "WARNING: TOWN is not NONE and you want LTOWN_TO_ROCK. "
-                    "Setting it to NONE!"
-                )
-            self.input_list.append(
-                {"json": {"NAM_PGD_ARRANGE_COVER": {"LTOWN_TO_ROCK": True}}}
-            )
-            self.input_list.append({"json": {"NAM_PGD_SCHEMES": {"TOWN": "NONE"}}})
-
-        if self.config.get_setting("SURFEX#TILES#INLAND_WATER") == "FLAKE":
-            self.input_list.append(
-                {
-                    "json": {
-                        "NAM_DATA_FLAKE": {
-                            "YWATER_DEPTH": "GlobalLakeDepth",
-                            "YWATER_DEPTHFILETYPE": "DIRECT",
-                            "YWATER_DEPTH_STATUS": "GlobalLakeStatus",
-                        }
-                    }
-                }
-            )
-
-        # Sea
-        self.input_list.append({"file": self.input_path + "/sea.json"})
-
-    def set_prep_namelist(
-        self, prep_file=None, prep_filetype=None, prep_pgdfile=None, prep_pgdfiletype=None
-    ):
-        """Set prep namelist.
-
-        Args:
-            prep_file (_type_, optional): _description_. Defaults to None.
-            prep_filetype (_type_, optional): _description_. Defaults to None.
-            prep_pgdfile (_type_, optional): _description_. Defaults to None.
-            prep_pgdfiletype (_type_, optional): _description_. Defaults to None.
-
-        Raises:
-            RuntimeError: Filetype for input to PREP is not set!
-            RuntimeError: Filetype for PGD input to PREP is not set!
-
-        """
-        if prep_file is not None and prep_filetype is None:
-            raise RuntimeError("Filetype for input to PREP is not set!")
-        if prep_pgdfile is not None and prep_pgdfiletype is None:
-            raise RuntimeError("Filetype for PGD input to PREP is not set!")
-
-        self.input_list.append({"file": self.input_path + "/prep.json"})
-        if prep_file is not None:
-            if prep_file.endswith(".json"):
-                self.input_list.append({"file": prep_file})
+        # Program specific settings
+        if self.program == "pgd":
+            eco_sg = self.config.get_setting("SURFEX#COVER#SG")
+            if eco_sg:
+                input_blocks += ["pgd_ecoclimap_sg"]
             else:
-                fname = os.path.basename(prep_file)
-                self.input_list.append({"json": {"NAM_PREP_SURF_ATM": {"CFILE": fname}}})
-                self.input_list.append(
-                    {"json": {"NAM_PREP_SURF_ATM": {"CFILETYPE": prep_filetype}}}
-                )
-                if prep_pgdfile is not None:
-                    fname = os.path.basename(prep_pgdfile)
-                    self.input_list.append(
-                        {"json": {"NAM_PREP_SURF_ATM": {"CFILEPGD": fname}}}
-                    )
-                    self.input_list.append(
-                        {
-                            "json": {
-                                "NAM_PREP_SURF_ATM": {"CFILEPGDTYPE": prep_pgdfiletype}
-                            }
-                        }
-                    )
-        if self.dtg is not None:
-            prep_time = self.dtg
-            self.input_list.append(
-                {"json": {"NAM_PREP_SURF_ATM": {"NYEAR": int(prep_time.strftime("%Y"))}}}
-            )
-            self.input_list.append(
-                {"json": {"NAM_PREP_SURF_ATM": {"NMONTH": int(prep_time.strftime("%m"))}}}
-            )
-            self.input_list.append(
-                {"json": {"NAM_PREP_SURF_ATM": {"NDAY": int(prep_time.strftime("%d"))}}}
-            )
-            self.input_list.append(
-                {
-                    "json": {
-                        "NAM_PREP_SURF_ATM": {
-                            "XTIME": float(prep_time.strftime("%H")) * 3600.0
-                        }
-                    }
-                }
-            )
-        else:
-            raise RuntimeError("You must provide a DTG for prep")
-        if self.config.get_setting("SURFEX#SEA#ICE") == "SICE":
-            self.input_list.append(
-                {"json": {"NAM_PREP_SEAFLUX": {"CSEAICE_SCHEME": "SICE"}}}
-            )
-            self.input_list.append({"file": self.input_path + "/prep_sice.json"})
+                input_blocks += ["pgd_ecoclimap"]
 
-        if self.config.get_setting("SURFEX#TILES#INLAND_WATER") == "FLAKE":
-            lclim_lake = self.config.get_setting("SURFEX#FLAKE#LCLIM")
-            self.input_list.append(
-                {"json": {"NAM_PREP_FLAKE": {"LCLIM_LAKE": lclim_lake}}}
-            )
+            # Set ISBA properties
+            if self.config.get_setting("SURFEX#ISBA#SCHEME") == "DIF":
+                input_blocks += ["pgd_isba_dif"]
+            elif self.config.get_setting("SURFEX#ISBA#SCHEME") == "3-L":
+                input_blocks += ["pgd_isba_3l"]
+            elif self.config.get_setting("SURFEX#ISBA#SCHEME") == "2-L":
+                input_blocks += ["pgd_isba_2l"]
 
-        # Set extra ISBA-DIF properties (Not needed in prep?)
-        if self.config.get_setting("SURFEX#ISBA#SCHEME") == "DIF":
-            if os.path.exists(self.input_path + "/isba_dif.json"):
-                self.input_list.append({"file": self.input_path + "/isba_dif.json"})
+            # Set MEB
+            if self.config.get_setting("SURFEX#ISBA#MEB"):
+                input_blocks += ["pgd_meb"]
 
-        # ISBA CANOPY
-        lisba_canopy = self.config.get_setting("SURFEX#ISBA#CANOPY")
-        self.input_list.append(
-            {"json": {"NAM_PREP_ISBA": {"LISBA_CANOPY": lisba_canopy}}}
-        )
-
-        # Snow
-        self.input_list.append({"file": self.input_path + "/prep_snow.json"})
-        if self.config.get_setting("SURFEX#ISBA#SNOW") == "D95":
-            self.input_list.append({"json": {"NAM_PREP_ISBA_SNOW": {"CSNOW": "D95"}}})
-        elif self.config.get_setting("SURFEX#ISBA#SNOW") == "3-L":
-            self.input_list.append({"json": {"NAM_PREP_ISBA_SNOW": {"CSNOW": "3-L"}}})
-        if self.config.get_setting("SURFEX#ISBA#SNOW") == "CRO":
-            self.input_list.append({"file": self.input_path + "/snow_crocus.json"})
-
-    def set_offline_namelist(self):
-        """Set offline namelist.
-
-        Raises:
-            RuntimeError: Mismatch in nnco/cobs_m
-
-        """
-        self.input_list.append({"file": self.input_path + "/offline.json"})
-
-        if self.config.get_setting("SURFEX#IO#LSELECT"):
-            self.input_list.append({"file": self.input_path + "/selected_output.json"})
-
-        # SEAFLX settings
-        if self.config.get_setting("SURFEX#TILES#SEA") == "SEAFLX":
-            # Surface perturbations
-            pertflux = self.config.get_setting("SURFEX#SEA#PERTFLUX")
-            self.input_list.append({"json": {"NAM_SEAFLUXn": {"LPERTFLUX": pertflux}}})
-
-        # ISBA settings
-        if self.config.get_setting("SURFEX#TILES#NATURE") == "ISBA":
-            pertsurf = self.config.get_setting("SURFEX#ISBA#PERTSURF")
-            self.input_list.append({"json": {"NAM_ISBAn": {"LPERTSURF": pertsurf}}})
-            xcgmax = self.config.get_setting("SURFEX#ISBA#XCGMAX", abort=False)
-            if xcgmax is not None:
-                self.input_list.append({"json": {"NAM_ISBAn": {"XCGMAX": xcgmax}}})
-            xcsmax = self.config.get_setting("SURFEX#ISBA#XCSMAX", abort=False)
-            if xcsmax is not None:
-                self.input_list.append({"json": {"NAM_ISBAn": {"XCSMAX": xcsmax}}})
-
-        # SSO
-        sso = self.config.get_setting("SURFEX#SSO#SCHEME")
-        self.input_list.append({"json": {"NAM_SSON": {"CROUGH": sso}}})
-        if sso == "OROTUR":
-            self.input_list.append({"json": {"NAM_SSON": {"XSOROT": self.geo.xdx}}})
-
-        # Perturbed offline settings
-        self.input_list.append({"json": {"NAM_VAR": {"NIVAR": 0}}})
-        self.input_list.append({"json": {"NAM_IO_VARASSIM": {"LPRT": False}}})
-        if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "EKF":
-            cassim_isba = self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA")
-            self.input_list.append({"json": {"NAM_ASSIM": {"CASSIM_ISBA": cassim_isba}}})
-            nvar = 0
-            cvar_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#CVAR_M")
-            nncv = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#NNCV")
-            xtprt_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XTPRT_M")
-            xsigma_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XSIGMA_M")
-            for var, cvar_val in enumerate(cvar_m):
-                self.input_list.append(
-                    {"json": {"NAM_VAR": {"CVAR_M(" + str(var + 1) + ")": cvar_val}}}
-                )
-                self.input_list.append(
-                    {"json": {"NAM_VAR": {"NNCV(" + str(var + 1) + ")": nncv[var]}}}
-                )
-                self.input_list.append(
-                    {"json": {"NAM_VAR": {"XTPRT_M(" + str(var + 1) + ")": xtprt_m[var]}}}
-                )
-                self.input_list.append(
-                    {
-                        "json": {
-                            "NAM_VAR": {"XSIGMA_M(" + str(var + 1) + ")": xsigma_m[var]}
-                        }
-                    }
-                )
-                if nncv[var] == 1:
-                    nvar += 1
-            self.input_list.append({"json": {"NAM_VAR": {"NVAR": nvar}}})
-
-        if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "ENKF":
-            self.input_list.append(
-                {
-                    "json": {
-                        "NAM_ASSIM": {
-                            "CASSIM_ISBA": self.config.get_setting(
-                                "SURFEX#ASSIM#SCHEMES#ISBA"
-                            )
-                        }
-                    }
-                }
-            )
-            nvar = 0
-            cvar_m = self.config.get_setting("SURFEX#ASSIM#ISBA#ENKF#CVAR_M")
-            nncv = self.config.get_setting("SURFEX#ASSIM#ISBA#ENKF#NNCV")
-            for var, cvar_val in enumerate(cvar_m):
-                self.input_list.append(
-                    {"json": {"NAM_VAR": {"CVAR_M(" + str(var + 1) + ")": cvar_val}}}
-                )
-                self.input_list.append(
-                    {"json": {"NAM_VAR": {"NNCV(" + str(var + 1) + ")": nncv[var]}}}
-                )
-                if nncv[var] == 1:
-                    nvar += 1
-            self.input_list.append({"json": {"NAM_VAR": {"NVAR": nvar}}})
-
-        # TODO the need for this in forecast must be removed!
-        nobstype = 0
-        nnco = self.config.get_setting("SURFEX#ASSIM#OBS#NNCO")
-        cobs_m = self.config.get_setting("SURFEX#ASSIM#OBS#COBS_M")
-        if len(nnco) != len(cobs_m):
-            raise RuntimeError("Mismatch in nnco/cobs_m")
-        for obs, obs_val in enumerate(nnco):
-            self.input_list.append(
-                {"json": {"NAM_OBS": {"NNCO(" + str(obs + 1) + ")": obs_val}}}
-            )
-            self.input_list.append(
-                {"json": {"NAM_OBS": {"COBS_M(" + str(obs + 1) + ")": cobs_m[obs]}}}
-            )
-            if nnco[obs] == 1:
-                nobstype += 1
-        self.input_list.append({"json": {"NAM_OBS": {"NOBSTYPE": nobstype}}})
-
-        # Climate setting
-        if self.config.get_setting("SURFEX#SEA#LVOLATILE_SIC"):
-            self.input_list.append(
-                {
-                    "json": {
-                        "NAM_SEAICEn ": {"LVOLATILE_SIC": True, "XSIC_EFOLDING_TIME": 1.0}
-                    }
-                }
-            )
-
-    def set_soda_namelist(self):
-        """Set SODA namelist.
-
-        Raises:
-            RuntimeError: Mismatch in nnco/cobs_m/xerrobs_m
-            RuntimeError: You must provide a DTG when using a list for snow
-            RuntimeError: Mismatch in nncv/cvar_m/xsigma_m/xtprt_m
-
-        """
-        self.input_list.append({"file": self.input_path + "/soda.json"})
-
-        self.input_list.append({"json": {"NAM_ASSIM": {"LASSIM": True}}})
-
-        lobsheader = self.config.get_setting("SURFEX#ASSIM#OBS#LOBSHEADER")
-        self.input_list.append({"json": {"NAM_OBS": {"LOBSHEADER": lobsheader}}})
-        lobsnat = self.config.get_setting("SURFEX#ASSIM#OBS#LOBSNAT")
-        self.input_list.append({"json": {"NAM_OBS": {"LOBSNAT": lobsnat}}})
-        cfile_format_obs = self.config.get_setting("SURFEX#ASSIM#OBS#CFILE_FORMAT_OBS")
-        self.input_list.append(
-            {"json": {"NAM_OBS": {"CFILE_FORMAT_OBS": cfile_format_obs}}}
-        )
-        nobstype = 0
-        nnco = self.config.get_setting("SURFEX#ASSIM#OBS#NNCO")
-        cobs_m = self.config.get_setting("SURFEX#ASSIM#OBS#COBS_M")
-        xerrobs_m = self.config.get_setting("SURFEX#ASSIM#OBS#XERROBS_M")
-        logging.debug("%s %s %s", nnco, cobs_m, xerrobs_m)
-        if len(nnco) != len(cobs_m) or len(nnco) != len(xerrobs_m):
-            raise RuntimeError("Mismatch in nnco/cobs_m/xerrobs_m")
-
-        for obs, obs_val in enumerate(nnco):
-            self.input_list.append(
-                {"json": {"NAM_OBS": {"NNCO(" + str(obs + 1) + ")": obs_val}}}
-            )
-            self.input_list.append(
-                {"json": {"NAM_OBS": {"COBS_M(" + str(obs + 1) + ")": cobs_m[obs]}}}
-            )
-            self.input_list.append(
-                {"json": {"NAM_OBS": {"XERROBS_M(" + str(obs + 1) + ")": xerrobs_m[obs]}}}
-            )
-            if nnco[obs] == 1:
-                nobstype += 1
-        self.input_list.append({"json": {"NAM_OBS": {"NOBSTYPE": nobstype}}})
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_OBS": {"LSWE": self.config.get_setting("SURFEX#ASSIM#OBS#LSWE")}
-                }
-            }
-        )
-
-        # LSM
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "CFILE_FORMAT_LSM": self.config.get_setting(
-                            "SURFEX#ASSIM#CFILE_FORMAT_LSM"
-                        )
-                    }
-                }
-            }
-        )
-
-        # Sea
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "CASSIM_SEA": self.config.get_setting("SURFEX#ASSIM#SCHEMES#SEA")
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "CFILE_FORMAT_SST": self.config.get_setting(
-                            "SURFEX#ASSIM#SEA#CFILE_FORMAT_SST"
-                        )
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "LREAD_SST_FROM_FILE": self.config.get_setting(
-                            "SURFEX#ASSIM#SEA#LREAD_SST_FROM_FILE"
-                        )
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "LEXTRAP_SEA": self.config.get_setting(
-                            "SURFEX#ASSIM#SEA#LEXTRAP_SEA"
-                        )
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "LECSST": self.config.get_setting("SURFEX#ASSIM#SEA#LECSST")
-                    }
-                }
-            }
-        )
-
-        # Water
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "CASSIM_WATER": self.config.get_setting(
-                            "SURFEX#ASSIM#SCHEMES#INLAND_WATER"
-                        )
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "LWATERTG2": self.config.get_setting(
-                            "SURFEX#ASSIM#INLAND_WATER#LWATERTG2"
-                        )
-                    }
-                }
-            }
-        )
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "LEXTRAP_WATER": self.config.get_setting(
-                            "SURFEX#ASSIM#INLAND_WATER#LEXTRAP_WATER"
-                        )
-                    }
-                }
-            }
-        )
-
-        # Nature
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "CASSIM_ISBA": self.config.get_setting(
-                            "SURFEX#ASSIM#SCHEMES#ISBA"
-                        )
-                    }
-                }
-            }
-        )
-
-        # Snow
-        self.input_list.append(
-            {
-                "json": {
-                    "NAM_ASSIM": {
-                        "LPATCH1": self.config.get_setting("SURFEX#ASSIM#ISBA#LPATCH1")
-                    }
-                }
-            }
-        )
-
-        laesnm = False
-        snow_cycles = self.config.get_setting("SURFEX#ASSIM#ISBA#UPDATE_SNOW_CYCLES")
-        if len(snow_cycles) > 0:
-            logging.debug("%s", self.dtg)
-            if self.dtg is not None:
-                for cycle in snow_cycles:
-                    logging.debug(self.dtg.strftime("%H"))
-                    if int(self.dtg.strftime("%H")) == int(cycle):
-                        logging.debug("true")
-                        laesnm = True
+            # RSMIN
+            if self.config.get_setting("SURFEX#COVER#SG"):
+                input_blocks += ["pgd_rsmin_sg"]
             else:
-                raise RuntimeError(
-                    "You must provide a DTG when using a list for snow  "
-                    "assimilation cycles"
-                )
-        self.input_list.append({"json": {"NAM_ASSIM": {"LAESNM": laesnm}}})
+                input_blocks += ["pgd_rsmin"]
 
-        # Set OI settings
-        if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "OI":
-            ua_physics = self.config.get_setting("FORECAST#PHYSICS", abort=False)
-            if ua_physics is None:
-                logging.info("Did not find FORECAST#PHYSICS. Assume arome")
-                ua_physics = "arome"
-            if ua_physics == "arome":
-                self.input_list.append({"json": {"NAM_ASSIM": {"LAROME": True}}})
-            elif ua_physics == "alaro":
-                self.input_list.append({"json": {"NAM_ASSIM": {"LAROME": False}}})
-
-            self.input_list.append(
-                {
-                    "json": {
-                        "NAM_NACVEG": {
-                            "XSIGT2MO": self.config.get_setting(
-                                "SURFEX#ASSIM#ISBA#OI#XSIGT2MO"
-                            )
-                        }
-                    }
-                }
-            )
-            self.input_list.append(
-                {
-                    "json": {
-                        "NAM_NACVEG": {
-                            "XSIGH2MO": self.config.get_setting(
-                                "SURFEX#ASSIM#ISBA#OI#XSIGH2MO"
-                            )
-                        }
-                    }
-                }
-            )
-            self.input_list.append({"json": {"NAM_NACVEG": {"XRCLIMCA": 0.0}}})
-            self.input_list.append({"json": {"NAM_NACVEG": {"XRCLISST": 0.05}}})
-            self.input_list.append({"json": {"NAM_NACVEG": {"NECHGU": self.fcint}}})
-            self.input_list.append({"json": {"NAM_NACVEG": {"LOBS2M": True}}})
-            self.input_list.append({"json": {"NAM_NACVEG": {"LOBSWG": False}}})
-            f_clim = self.config.get_setting("SURFEX#ASSIM#ISBA#OI#CFILE_FORMAT_CLIM")
-            self.input_list.append({"json": {"NAM_ASSIM": {"CFILE_FORMAT_CLIM": f_clim}}})
-            f_fg = self.config.get_setting("SURFEX#ASSIM#ISBA#OI#CFILE_FORMAT_FG")
-            self.input_list.append({"json": {"NAM_ASSIM": {"CFILE_FORMAT_FG": f_fg}}})
-
-        if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "EKF":
-            nvar = 0
-            llincheck = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#LLINCHECK")
-            self.input_list.append({"json": {"NAM_ASSIM": {"LLINCHECK": llincheck}}})
-            xalpha = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XALPHA")
-            self.input_list.append({"json": {"NAM_VAR": {"XALPHA": xalpha}}})
-            cvar_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#CVAR_M")
-            xsigma_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XSIGMA_M")
-            xtprt_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XTPRT_M")
-            nncv = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#NNCV")
-            if (
-                len(nncv) != len(cvar_m)
-                or len(nncv) != len(xsigma_m)
-                or len(nncv) != len(xtprt_m)
-            ):
-                raise RuntimeError("Mismatch in nncv/cvar_m/xsigma_m/xtprt_m")
-            for var, cvar_val in enumerate(cvar_m):
-                self.input_list.append(
-                    {"json": {"NAM_VAR": {"CVAR_M(" + str(var + 1) + ")": cvar_val}}}
-                )
-                self.input_list.append(
-                    {
-                        "json": {
-                            "NAM_VAR": {"XSIGMA_M(" + str(var + 1) + ")": xsigma_m[var]}
-                        }
-                    }
-                )
-                self.input_list.append(
-                    {"json": {"NAM_VAR": {"XTPRT_M(" + str(var + 1) + ")": xtprt_m[var]}}}
-                )
-                self.input_list.append(
-                    {"json": {"NAM_VAR": {"NNCV(" + str(var + 1) + ")": nncv[var]}}}
-                )
-                if nncv[var] == 1:
-                    nvar += 1
-            self.input_list.append({"json": {"NAM_VAR": {"NIVAR": 0}}})
-            self.input_list.append({"json": {"NAM_VAR": {"NVAR": nvar}}})
-            xscale_q = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XSCALE_Q")
-            self.input_list.append({"json": {"NAM_VAR": {"XSCALE_Q": xscale_q}}})
-            self.input_list.append(
-                {
-                    "json": {
-                        "NAM_IO_VARASSIM": {
-                            "LPRT": False,
-                            "LBEV": self.config.get_setting(
-                                "SURFEX#ASSIM#ISBA#EKF#EVOLVE_B"
-                            ),
-                            "LBFIXED": not self.config.get_setting(
-                                "SURFEX#ASSIM#ISBA#EKF#EVOLVE_B"
-                            ),
-                        }
-                    }
-                }
-            )
-
-        if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "ENKF":
-            nvar = 0
-            cvar_m = self.config.get_setting("SURFEX#ASSIM#ISBA#ENKF#CVAR_M")
-            nncv = self.config.get_setting("SURFEX#ASSIM#ISBA#ENKF#NNCV")
-            if len(nncv) != len(cvar_m):
-                raise RuntimeError("Mismatch in nncv/cvar_m")
-            for var, cvar_val in enumerate(cvar_m):
-                self.input_list.append(
-                    {"json": {"NAM_VAR": {"CVAR_M(" + str(var + 1) + ")": cvar_val}}}
-                )
-                self.input_list.append(
-                    {"json": {"NAM_VAR": {"NNCV(" + str(var + 1) + ")": nncv[var]}}}
-                )
-
-                if nncv[var] == 1:
-                    nvar += 1
-            self.input_list.append({"json": {"NAM_VAR": {"NIVAR": 0}}})
-            self.input_list.append({"json": {"NAM_VAR": {"NVAR": nvar}}})
-
-        # Town
-        town_setting = self.config.get_setting("SURFEX#ASSIM#SCHEMES#TEB")
-        self.input_list.append({"json": {"NAM_ASSIM": {"CASSIM_TEB": town_setting}}})
-
-    def epilog(self):
-        """Epilog."""
-        # Always set these
-        if self.config.get_setting("SURFEX#SEA#ICE") == "SICE":
-            self.input_list.append({"file": self.input_path + "/sice.json"})
-
-        self.input_list.append({"file": self.input_path + "/treedrag.json"})
-        lfaketrees = self.config.get_setting("SURFEX#TREEDRAG#FAKETREES", abort=False)
-        if lfaketrees is not None:
-            self.input_list.append({"json": {"NAM_TREEDRAG": {"LFAKETREE": lfaketrees}}})
-
-        if self.config.get_setting("SURFEX#TILES#INLAND_WATER") == "FLAKE":
-            self.input_list.append({"file": self.input_path + "/flake.json"})
-
-    def override(self):
-        """Overide."""
-        # Override posssibility
-        if os.path.exists(self.input_path + "/override.json"):
-            logging.warning(
-                "WARNING: Override settings with content from %s/override.json",
-                self.input_path,
-            )
-            self.input_list.append({"file": self.input_path + "/override.json"})
-
-    @staticmethod
-    def set_direct_data_namelist(lnamelist_section, ldtype, ldname, linput_path):
-        """Set direct data namelist.
-
-        Args:
-            lnamelist_section (_type_): _description_
-            ldtype (_type_): _description_
-            ldname (_type_): _description_
-            linput_path (_type_): _description_
-
-        Returns:
-            _type_: _description_
-
-        """
-        if ldname.endswith(".dir"):
-            basename = os.path.splitext(os.path.basename(ldname))[0]
-            filetype_name = ldtype
-            if ldtype == "YSOC_TOP" or ldtype == "YSOC_SUB":
-                filetype_name = "YSOC"
-            return {
-                "json": json.loads(
-                    '{"'
-                    + lnamelist_section
-                    + '": { "'
-                    + ldtype
-                    + '": "'
-                    + basename
-                    + '", '
-                    + '"'
-                    + filetype_name
-                    + 'FILETYPE": "DIRECT"}}'
-                )
-            }
-        if ldname.endswith(".json"):
-            return {"file": linput_path + "/" + ldname}
-
-    @staticmethod
-    def set_dirtyp_data_namelist(
-        lnamelist_section, ldtype, ldname, vtype=None, decade=None
-    ):
-        """Set dirtyp data namelist.
-
-        Args:
-            lnamelist_section (_type_): _description_
-            ldtype (_type_): _description_
-            ldname (_type_): _description_
-            vtype (_type_, optional): _description_. Defaults to None.
-            decade (_type_, optional): _description_. Defaults to None.
-
-        Returns:
-            _type_: _description_
-
-        """
-        basename = os.path.splitext(os.path.basename(ldname))[0]
-        filetype_name = ldtype.upper()
-        if vtype is not None or decade is not None:
-            filetype_name = filetype_name + "("
-        if vtype is not None:
-            filetype_name = filetype_name + str(vtype)
-        if vtype is not None and decade is not None:
-            filetype_name = filetype_name + ","
-        if decade is not None:
-            filetype_name = filetype_name + str(decade)
-        if vtype is not None or decade is not None:
-            filetype_name = filetype_name + ")"
-        return {
-            "json": json.loads(
-                '{"'
-                + lnamelist_section
-                + '": { "CFNAM_'
-                + filetype_name
-                + '": "'
-                + basename
-                + '", "CFTYP_'
-                + filetype_name
-                + '": "DIRTYPE"}}'
-            )
-        }
-
-    @staticmethod
-    def capitalize_namelist_dict(dict_in):
-        """Capitalize namelist.
-
-        Args:
-            dict_in (dict): _description_
-
-        Returns:
-            _type_: _description_
-
-        """
-        new_dict = {}
-        for key in dict_in:
-            upper_case2 = {}
-            for key2 in dict_in[key]:
-                upper_case2.update({key2.upper(): dict_in[key][key2]})
-            new_dict.update({key.upper(): upper_case2})
-        return new_dict
-
-    @staticmethod
-    def lower_case_namelist_dict(dict_in):
-        """Lower case namelist.
-
-        Args:
-            dict_in (dict): Namelist dictionary to lower case
-
-        Returns:
-            dict: Namelist in lower case
-
-        """
-        new_dict = {}
-        for key in dict_in:
-            lower_case_dict = {}
-            for key2 in dict_in[key]:
-                lower_case_dict.update({key2.lower(): dict_in[key][key2]})
-            new_dict.update({key.lower(): lower_case_dict})
-        return new_dict
-
-    @staticmethod
-    def merge_namelist_dicts(old_dict, new_dict):
-        """Merge namelist dicts.
-
-        Args:
-            old_dict (dict): Old dictionary
-            new_dict (dict): New dictionary
-
-        Returns:
-            dict: Merged dictionary
-
-        """
-        old_dict = Namelist.capitalize_namelist_dict(old_dict)
-        new_dict = Namelist.capitalize_namelist_dict(new_dict)
-        merged_dict = old_dict
-
-        for new_key, new_val in new_dict.items():
-            # Namelist block already exists
-            if new_key in merged_dict:
-                settings = merged_dict[new_key]
-                for new_key2 in new_val:
-                    settings.update({new_key2: new_dict[new_key][new_key2]})
-
-                merged_dict.update({new_key: settings})
-            # New namelist block
+            # CV
+            if self.config.get_setting("SURFEX#COVER#SG"):
+                input_blocks += ["pgd_cv_sg"]
             else:
-                merged_dict.update({new_key: new_val})
+                input_blocks += ["pgd_cv"]
 
-        return merged_dict
+            # Treedrag
+            input_blocks += ["pgd_treedrag"]
 
-    @staticmethod
-    def ascii2nml(input_data):
-        """Convert dict to a namelist object.
+            if self.config.get_setting("SURFEX#TOWN#LTOWN_TO_ROCK"):
+                input_blocks += ["pgd_town_to_rock"]
 
-        Args:
-            input_data (dict): Namelist settings
+            if self.config.get_setting("SURFEX#TILES#INLAND_WATER") == "FLAKE":
+                input_blocks += ["pgd_flake"]
 
-        Returns:
-            f90nml.Namelist: Namelist object.
-        """
-        output_data = f90nml.Namelist(input_data)
-        return output_data
+            # Sea
+            input_blocks += ["pgd_sea"]
 
-    @staticmethod
-    def ascii_file2nml(input_fname, input_fmt="json"):
-        """Convert a file wih namelist settings to a Namelist object.
+        elif self.program == "prep":
+            input_blocks += ["prep"]
 
-        Args:
-            input_fname (str): Filname
-            input_fmt (str, optional): File format. Defaults to "json".
+            if self.config.get_setting("SURFEX#SEA#ICE") == "SICE":
+                input_blocks += ["prep_sice"]
 
-        Returns:
-            f90nml.Namelist: Namelist object.
+            if self.config.get_setting("SURFEX#TILES#INLAND_WATER") == "FLAKE":
+                input_blocks += ["prep_flake"]
 
-        """
-        if input_fmt == "json":
-            with open(input_fname, mode="r", encoding="utf-8") as input_file:
-                output_data = json.load(input_file)
-        elif input_fmt == "yaml":
-            with open(input_fname, mode="r", encoding="utf-8") as input_file:
-                output_data = yaml.safe_load(input_file)
-        output_data = f90nml.Namelist(output_data)
-        return output_data
+            # ISBA
+            if self.config.get_setting("SURFEX#ISBA#SCHEME") == "DIF":
+                input_blocks += ["prep_isba", "prep_isba_dif"]
+            elif self.config.get_setting("SURFEX#ISBA#SCHEME") == "3-L":
+                input_blocks += ["prep_isba", "prep_isba_3l"]
+            elif self.config.get_setting("SURFEX#ISBA#SCHEME") == "2-L":
+                input_blocks += ["prep_isba", "prep_isba_2l"]
 
-    @staticmethod
-    def nml2ascii(input_data, output_file, output_fmt="json", indent=2):
-        """Dump a namelist object as a dict in a json or yaml file.
+            if lisba:
+                # ISBA CANOPY
+                if self.config.get_setting("SURFEX#ISBA#CANOPY"):
+                    input_blocks += ["prep_isba_canopy"]
 
-        Args:
-            input_data (f90nml.Namelist): Namelist object.
-            output_file (str): Filename
-            output_fmt (str, optional): File format. Defaults to "json".
-            indent (int, optional): Indentation. Defaults to 2.
+                # Snow
+                input_blocks += ["prep_isba_snow"]
+                if self.config.get_setting("SURFEX#ISBA#SNOW") == "D95":
+                    input_blocks += ["prep_isba_snow_d95"]
+                elif self.config.get_setting("SURFEX#ISBA#SNOW") == "3-L":
+                    input_blocks += ["prep_isba_snow_3l"]
+                if self.config.get_setting("SURFEX#ISBA#SNOW") == "CRO":
+                    input_blocks += ["prep_isba_snow_cro"]
 
-        """
-        if output_fmt == "json":
-            input_data = input_data.todict(complex_tuple=True)
-            json.dump(
-                input_data,
-                open(output_file, "w", encoding="utf-8"),
-                indent=indent,
-                separators=(",", ": "),
-            )
-        elif output_fmt == "yaml":
-            input_data = input_data.todict(complex_tuple=True)
-            yaml.dump(input_data, output_file, default_flow_style=False)
+        elif self.program == "offline" or self.program == "perturbed":
+            input_blocks += ["offline"]
+            input_blocks += ["offline_selected_output"]
 
-    @staticmethod
-    def merge_json_namelist_file(old_dict, my_file):
-        """Merge json files with namelists.
+            # SEAFLX settings
+            if self.config.get_setting("SURFEX#TILES#SEA") == "SEAFLX":
+                input_blocks += ["offline_seaflux"]
 
-        Args:
-            old_dict (dict): Exististing settings
-            my_file (str): Filename with new settings
+            # ISBA settings
+            if lisba:
+                input_blocks += ["offline_isba"]
+                if self.config.get_setting("SURFEX#ISBA#PERTSURF"):
+                    input_blocks += ["offline_isba_pertsurf"]
 
-        Raises:
-            FileNotFoundError: Namelist input not found
+            # SSO
+            sso = self.config.get_setting("SURFEX#SSO#SCHEME").lower()
+            input_blocks += ["offline_" + sso]
 
-        Returns:
-            dict: Merged settings.
+            # Perturbed offline settings
+            if lisba:
+                input_blocks += ["offline_pert_isba_settings"]
+                if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "EKF":
+                    input_blocks += ["offline_pert_isba_ekf"]
 
-        """
-        logging.debug(my_file)
-        if os.path.exists(my_file):
-            with open(my_file, "r", encoding="utf-8") as file_handler:
-                new_dict = json.load(file_handler)
+                if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "ENKF":
+                    input_blocks += ["offline_pert_isba_enkf"]
+
+            input_blocks += ["offline_pert_obs"]
+
+            # Climate setting
+            if self.config.get_setting("SURFEX#SEA#LVOLATILE_SIC"):
+                input_blocks += ["offline_volatile_sic"]
+
+        elif self.program == "soda":
+            input_blocks += ["soda", "soda_obs"]
+
+            # Set OI settings
+            if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "OI":
+                input_blocks += ["soda_isba_oi"]
+
+            elif self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "EKF":
+                input_blocks += ["soda_isba_ekf"]
+
+            elif self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "ENKF":
+                input_blocks += ["soda_isba_enkf"]
+
+            # Town
+            if self.config.get_setting("SURFEX#ASSIM#SCHEMES#TEB").lower() != "none":
+                input_blocks += ["soda_teb"]
+
         else:
-            raise FileNotFoundError(my_file)
+            raise NotImplementedError(self.program)
 
-        return Namelist.merge_namelist_dicts(old_dict, new_dict)
+        return input_blocks
 
     def get_namelist(self):
         """Get namelist."""
-        logging.debug("Constructing namelist:")
-        merged_json_settings = {}
-        for inp in self.input_list:
-            if "file" in inp:
-                json_file = str(inp["file"])
-                if not os.path.exists(json_file):
-                    raise FileNotFoundError(
-                        "Needed namelist input does not exist: " + json_file
-                    )
-                else:
-                    merged_json_settings = self.merge_json_namelist_file(
-                        merged_json_settings, json_file
-                    )
-            elif "json" in inp:
-                merged_json_settings = self.merge_namelist_dicts(
-                    merged_json_settings, inp["json"]
-                )
+        return self.nml
+
+    @staticmethod
+    def check_nml_setting(problems, nml, block, key, value):
+
+        ckey = block + "." + key
+        if block in nml:
+            if key in nml[block]:
+                if nml[block][key] != value:
+                    msg = f"Mismatch: {nml[block][key]} != {value}"
+                    problems.update({ckey: {"SEVERE": msg}})
             else:
-                logging.error("Can not handle input type %s", str(inp))
-                raise Exception
+                problems.update({ckey: {"WARNING": "Namelist key not found"}})
+        else:
+            problems.update({ckey: {"WARNING": "Namelist block not found"}})
+        return problems
 
-        return self.ascii2nml(merged_json_settings)
-
-
-class Namelist(object):
-    """Base namelist."""
-
-    def __init__(
-        self,
-        program,
-        config,
-        input_path,
-        forc_zs=False,
-        prep_file=None,
-        prep_filetype=None,
-        prep_pgdfile=None,
-        prep_pgdfiletype=None,
-        dtg=None,
-        fcint=3,
-        geo=None,
-    ):
-        """Construct a base namelists class to be implemented by namelist implementations.
+    def concistency(self, nml, abort=True):
+        """Check if namelist is consistent with config
 
         Args:
-            program (str): Which surfex binary you want to run ["pgd", "prep", "offline", "soda"]
-            config (surfex.Configuration): A SURFEX configuration object
-            input_path (str): A path to search for json namelist blocks
-            forc_zs (bool): Set the surfex orography heigth to the same as in the forcing files
-            prep_file (str): Input file for prep
-            prep_filetype (str): The format of prep_file
-            prep_pgdfile (str): Input PGD file for prep in case it comes from SURFEX
-            prep_pgdfiletype (str): The format of the prep_pgdfile
-            dtg (datetime.datetime): The date/time you want to run
-            fcint (int): The intervall between the cycles. Used for first guesses.
-            geo (surfex.Geo): Surfex geometry. The domain you want to run on
+            nml (f90nml.Namelist): A parsed f90nml namelist
+            abort (bool, optional): Abort if mismatch. Defaults to True.
+
 
         Raises:
             RuntimeError: Input
@@ -1229,1125 +243,176 @@ class Namelist(object):
             NotImplementedError: Mode is not implemented
 
         """
-        self.config = config
-        self.input_path = input_path
-        self.forc_zs = forc_zs
-        if dtg is not None:
-            if isinstance(dtg, str):
-                dtg = as_datetime(dtg)
-        self.dtg = dtg
-        check_parsing = True
-        if self.dtg is None:
-            check_parsing = False
-        self.fcint = fcint
-        self.geo = geo
 
-        # The time stamp of next cycle file
-        forecast_length = self.fcint
-        if self.dtg is not None:
-            self.end_of_forecast = self.dtg + as_timedelta(seconds=forecast_length * 3600)
-        else:
-            self.end_of_forecast = None
+        logging.info("Check namelist input for program: %s", self.program)
 
-        logging.info("Creating JSON namelist input for program: %s", program)
-
-        merged_dict = {}
-        merged_dict = self.prolog(merged_dict, check_parsing=check_parsing)
-        # Program specific settings
-        if program == "pgd":
-            merged_dict = self.set_pgd_namelist(merged_dict)
-        elif program == "prep":
-            if prep_file is None:
-                raise RuntimeError(
-                    "Prep need an input file either as a json namelist or a surfex "
-                    "supported format"
-                )
-            merged_dict = self.set_prep_namelist(
-                merged_dict,
-                prep_file=prep_file,
-                prep_filetype=prep_filetype,
-                prep_pgdfile=prep_pgdfile,
-                prep_pgdfiletype=prep_pgdfiletype,
-            )
-        elif program == "offline" or program == "perturbed":
-            merged_dict = self.set_offline_namelist(merged_dict)
-        elif program == "soda":
-            merged_dict = self.set_soda_namelist(merged_dict)
-        else:
-            raise NotImplementedError(program)
-        merged_dict = self.epilog(merged_dict)
-        merged_dict = self.override(merged_dict)
-
-        # Check parsing
-        non_parsed = {}
-        for block in merged_dict:
-            keys = []
-            for key, value in merged_dict[block].items():
-                if isinstance(value, str) and value[0] == "@" and value[-1] == "@":
-                    logging.info(
-                        "Delete non-substituted placeholder %s for key %s", value, key
-                    )
-                    keys.append(key)
-            if len(keys) > 0:
-                non_parsed.update({block: keys})
-
-        for block, keys in non_parsed.items():
-            for key in keys:
-                merged_dict = self.delete(merged_dict, block, key)
-
-        for block in merged_dict:
-            for key, value in merged_dict[block].items():
-                if isinstance(value, str) and "@" in value[0] and "@" in value[-1]:
-                    logging.error(
-                        "Merged dictionary contains a @ in value %s for key %s",
-                        value,
-                        key,
-                    )
-                    raise RuntimeError("Merged dictionary contains a @ in value")
-
-        self.namelist_dict = merged_dict
-
-    def prolog(self, merged_dict, check_parsing=True):
-        """Prolog.
-
-        Args:
-            merged_dict(dict): Merged settings.
-            check_parsing (bool, optional): Check if parsing is ok. Defaults to True.
-
-        Returns:
-            merged_dict(dict): Merged settings
-
-        """
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/prolog.json"
-        )
-        # IO
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/io.json"
-        )
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/constants.json"
-        )
-
-        # IO manipulation
-        merged_dict = self.sub(
-            merged_dict,
+        problems = {}
+        problems = self.check_nml_setting(
+            problems,
+            nml,
             "NAM_IO_OFFLINE",
             "CSURF_FILETYPE",
             self.config.get_setting("SURFEX#IO#CSURF_FILETYPE"),
         )
-        merged_dict = self.sub(
-            merged_dict,
+        problems = self.check_nml_setting(
+            problems,
+            nml,
             "NAM_IO_OFFLINE",
             "CTIMESERIES_FILETYPE",
             self.config.get_setting("SURFEX#IO#CTIMESERIES_FILETYPE"),
         )
-        merged_dict = self.sub(
-            merged_dict,
+        problems = self.check_nml_setting(
+            problems,
+            nml,
             "NAM_IO_OFFLINE",
             "CFORCING_FILETYPE",
             self.config.get_setting("SURFEX#IO#CFORCING_FILETYPE"),
         )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_IO_OFFLINE",
-            "CPGDFILE",
-            self.config.get_setting("SURFEX#IO#CPGDFILE"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_IO_OFFLINE",
-            "CPREPFILE",
-            self.config.get_setting("SURFEX#IO#CPREPFILE"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_IO_OFFLINE",
-            "CSURFFILE",
-            self.config.get_setting(
-                "SURFEX#IO#CSURFFILE",
-                validtime=self.end_of_forecast,
-                basedtg=self.dtg,
-                check_parsing=check_parsing,
-            ),
-        )
-        merged_dict = self.sub(
-            merged_dict,
+
+        problems = self.check_nml_setting(
+            problems,
+            nml,
             "NAM_IO_OFFLINE",
             "XTSTEP_SURF",
             self.config.get_setting("SURFEX#IO#XTSTEP"),
         )
-        merged_dict = self.sub(
-            merged_dict,
+        problems = self.check_nml_setting(
+            problems,
+            nml,
             "NAM_IO_OFFLINE",
             "XTSTEP_OUTPUT",
             self.config.get_setting("SURFEX#IO#XTSTEP_OUTPUT"),
         )
-        merged_dict = self.sub(
-            merged_dict,
+        problems = self.check_nml_setting(
+            problems,
+            nml,
             "NAM_WRITE_SURF_ATM",
             "LSPLIT_PATCH",
             self.config.get_setting("SURFEX#IO#LSPLIT_PATCH"),
         )
-        merged_dict = self.sub(
-            merged_dict, "NAM_IO_OFFLINE", "LSET_FORC_ZS", self.forc_zs
-        )
 
         # Constants and parameters
-        merged_dict = self.sub(
-            merged_dict,
+        problems = self.check_nml_setting(
+            problems,
+            nml,
             "NAM_SURF_ATM",
             "XRIMAX",
             self.config.get_setting("SURFEX#PARAMETERS#XRIMAX"),
         )
-        return merged_dict
 
-    def set_pgd_namelist(self, merged_dict):
-        """Set pgd namelist."""
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/pgd.json"
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_PGD_SCHEMES",
-            "CSEA",
-            self.config.get_setting("SURFEX#TILES#SEA"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_PGD_SCHEMES",
-            "CWATER",
-            self.config.get_setting("SURFEX#TILES#INLAND_WATER"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_PGD_SCHEMES",
-            "CNATURE",
-            self.config.get_setting("SURFEX#TILES#NATURE"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_PGD_SCHEMES",
-            "CTOWN",
-            self.config.get_setting("SURFEX#TILES#TOWN"),
-        )
-        if self.config.get_setting("SURFEX#TOWN#LTOWN_TO_ROCK"):
+        # Program specific settings
+        if self.program == "pgd":
 
-            if self.config.get_setting("SURFEX#TILES#TOWN") != "NONE":
-                logging.warning(
-                    "WARNING: TOWN is not NONE and you want LTOWN_TO_ROCK. "
-                    "Setting it to NONE!"
-                )
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_arrange_cover.json"
+            problems = self.check_nml_setting(
+                problems,
+                nml,
+                "NAM_PGD_SCHEMES",
+                "CSEA",
+                self.config.get_setting("SURFEX#TILES#SEA"),
             )
-            merged_dict = self.sub(
-                merged_dict, "NAM_PGD_ARRANGE_COVER", "LTOWN_TO_ROCK", True
+            problems = self.check_nml_setting(
+                problems,
+                nml,
+                "NAM_PGD_SCHEMES",
+                "CWATER",
+                self.config.get_setting("SURFEX#TILES#INLAND_WATER"),
             )
-            merged_dict = self.sub(merged_dict, "NAM_PGD_SCHEMES", "CTOWN", "NONE")
-
-        # Cover
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/pgd_cover.json"
-        )
-
-        # COVER
-        eco_sg = self.config.get_setting("SURFEX#COVER#SG")
-        merged_dict = self.sub(merged_dict, "NAM_FRAC", "LECOSG", eco_sg)
-        # Ecoclimap SG
-        if eco_sg:
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_eco_sg.json"
+            problems = self.check_nml_setting(
+                problems,
+                nml,
+                "NAM_PGD_SCHEMES",
+                "CNATURE",
+                self.config.get_setting("SURFEX#TILES#NATURE"),
             )
-            ecoclimap = EcoclimapSG(self.config)
-        else:
-            ecoclimap = Ecoclimap(self.config)
-
-        fname = str(self.config.get_setting("SURFEX#COVER#YCOVER"))
-        fname, filetype = self.get_filetype_from_suffix(fname)
-        merged_dict = self.sub(merged_dict, "NAM_COVER", "YCOVER", fname)
-        if filetype is not None:
-            merged_dict = self.sub(merged_dict, "NAM_COVER", "YCOVERFILETYPE", filetype)
-
-        # ZS
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/pgd_zs.json"
-        )
-        fname = str(self.config.get_setting("SURFEX#ZS#YZS"))
-        fname, filetype = self.get_filetype_from_suffix(fname)
-        merged_dict = self.sub(merged_dict, "NAM_ZS", "YZS", fname)
-        if filetype is not None:
-            merged_dict = self.sub(merged_dict, "NAM_ZS", "YZSFILETYPE", filetype)
-
-        # Sea
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/pgd_sea.json"
-        )
-
-        # Inland water
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/pgd_inland_water.json"
-        )
-
-        if self.config.get_setting("SURFEX#TILES#INLAND_WATER") == "FLAKE":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_flake.json"
+            problems = self.check_nml_setting(
+                problems,
+                nml,
+                "NAM_PGD_SCHEMES",
+                "CTOWN",
+                self.config.get_setting("SURFEX#TILES#TOWN"),
             )
+            if self.config.get_setting("SURFEX#TOWN#LTOWN_TO_ROCK"):
 
-        # PGD ISBA
-        if self.config.get_setting("SURFEX#TILES#NATURE") == "ISBA":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_isba.json"
-            )
-
-        if eco_sg:
-            merged_dict = self.sub(
-                merged_dict, "NAM_DATA_ISBA", "NTIME", ecoclimap.decades
-            )
-            fname = self.config.get_setting("SURFEX#COVER#H_TREE")
-            if fname != "" and fname is not None:
-                if fname.endswith(".dir"):
-                    fname = fname.replace(".dir", "")
-                for vtt in range(1, ecoclimap.veg_types + 1):
-                    merged_dict = self.sub(
-                        merged_dict,
-                        "NAM_DATA_ISBA",
-                        "CFNAM_H_TREE(@VEGTYPE@)",
-                        fname,
-                        vtype=vtt,
-                    )
-                    merged_dict = self.sub(
-                        merged_dict,
-                        "NAM_DATA_ISBA",
-                        "CFTYP_H_TREE(@VEGTYPE@)",
-                        "DIRTYPE",
-                        vtype=vtt,
+                if self.config.get_setting("SURFEX#TILES#TOWN") != "NONE":
+                    logging.warning(
+                        "WARNING: TOWN is not NONE and you want LTOWN_TO_ROCK. "
+                        "Setting it to NONE!"
                     )
 
-                key = "CFNAM_H_TREE(@VEGTYPE@)"
-                merged_dict = self.delete(merged_dict, "NAM_DATA_ISBA", key)
-                key = "CFTYP_H_TREE(@VEGTYPE@)"
-                merged_dict = self.delete(merged_dict, "NAM_DATA_ISBA", key)
-
-            decadal_data_types = [
-                "ALBNIR_SOIL",
-                "ALBNIR_VEG",
-                "ALBVIS_SOIL",
-                "ALBVIS_VEG",
-                "LAI",
-            ]
-            for decadal_data_type in decadal_data_types:
-                for vtt in range(1, ecoclimap.veg_types + 1):
-                    for decade in range(1, ecoclimap.decades + 1):
-                        filepattern = self.config.get_setting(
-                            "SURFEX#COVER#" + decadal_data_type, check_parsing=False
-                        )
-                        fname = ecoclimap.parse_fnames(filepattern, decade)
-                        if fname.endswith(".dir"):
-                            fname = fname.replace(".dir", "")
-                        key = f"CFNAM_{decadal_data_type}(@VEGTYPE@,@DECADE@)"
-                        merged_dict = self.sub(
-                            merged_dict,
-                            "NAM_DATA_ISBA",
-                            key,
-                            fname,
-                            vtype=vtt,
-                            decade=decade,
-                        )
-                        key = f"CFTYP_{decadal_data_type}(@VEGTYPE@,@DECADE@)"
-                        merged_dict = self.sub(
-                            merged_dict,
-                            "NAM_DATA_ISBA",
-                            key,
-                            "DIRTYPE",
-                            vtype=vtt,
-                            decade=decade,
-                        )
-
-                # Delete generic placehoder
-                key = f"CFNAM_{decadal_data_type}(@VEGTYPE@,@DECADE@)"
-                merged_dict = self.delete(merged_dict, "NAM_DATA_ISBA", key)
-                key = f"CFTYP_{decadal_data_type}(@VEGTYPE@,@DECADE@)"
-                merged_dict = self.delete(merged_dict, "NAM_DATA_ISBA", key)
-
-        isba_input_data = ["YSAND", "YCLAY", "YSOC_TOP", "YSOC_SUB"]
-        for ftype in isba_input_data:
-            fname = str(self.config.get_setting("SURFEX#ISBA#" + ftype))
-            fname, filetype = self.get_filetype_from_suffix(fname)
-            merged_dict = self.sub(merged_dict, "NAM_ISBA", ftype, fname)
-            key = ftype + "FILETYPE"
-            if ftype in ["YSOC_TOP", "YSOC_SUB"]:
-                if ftype in ["YSOC_TOP"]:
-                    key = "YSOCFILETYPE"
-                    merged_dict = self.sub(merged_dict, "NAM_ISBA", key, filetype)
-            else:
-                merged_dict = self.sub(merged_dict, "NAM_ISBA", key, filetype)
-
-        # Set ISBA properties
-        if self.config.get_setting("SURFEX#ISBA#SCHEME") == "DIF":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_isba_dif.json"
-            )
-            merged_dict = self.sub(merged_dict, "NAM_ISBA", "CISBA", "DIF")
-            merged_dict = self.sub(merged_dict, "NAM_ISBA", "NGROUND_LAYER", 14)
-        elif self.config.get_setting("SURFEX#ISBA#SCHEME") == "3-L":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_isba_3-L.json"
-            )
-            merged_dict = self.sub(merged_dict, "NAM_ISBA", "CISBA", "3-L")
-            merged_dict = self.sub(merged_dict, "NAM_ISBA", "NGROUND_LAYER", 3)
-        elif self.config.get_setting("SURFEX#ISBA#SCHEME") == "2-L":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_isba_2-L.json"
-            )
-            merged_dict = self.sub(merged_dict, "NAM_ISBA", "CISBA", "2-L")
-            merged_dict = self.sub(merged_dict, "NAM_ISBA", "NGROUND_LAYER", 2)
-
-        # Set patches
-        npatch = self.config.get_setting("SURFEX#ISBA#NPATCH")
-        merged_dict = self.sub(merged_dict, "NAM_ISBA", "NPATCH", npatch)
-
-        # Set MEB
-        lmeb = self.config.get_setting("SURFEX#ISBA#MEB")
-        merged_dict = self.sub(merged_dict, "NAM_ISBA", "LMEB", lmeb)
-        if lmeb:
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/meb_settings.json"
+            problems = self.check_nml_setting(
+                problems,
+                nml,
+                "NAM_FRAC",
+                "LECOSG",
+                self.config.get_setting("SURFEX#COVER#SG"),
             )
 
-        # RSMIN
-        if self.config.get_setting("SURFEX#COVER#SG"):
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_rsmin_sg.json"
-            )
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_rsmin_sg_mod.json"
-            )
-        else:
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_rsmin.json"
-            )
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_rsmin_mod.json"
-            )
-
-        # CV
-        if self.config.get_setting("SURFEX#COVER#SG"):
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_cv_sg.json"
-            )
-        else:
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/pgd_cv.json"
-            )
-
-        # Treedrag (from ASCLLV files)
-        if self.config.get_setting("SURFEX#TREEDRAG#TREEDATA_FILE") != "":
-            treeheight = self.config.get_setting("SURFEX#TREEDRAG#TREEDATA_FILE")
-            for vtt in [4, 5, 6]:
-                merged_dict = self.sub(
-                    merged_dict,
-                    "NAM_DATA_ISBA",
-                    "CFNAM_H_TREE(@VEGTYPE@)",
-                    treeheight,
-                    vtype=vtt,
-                )
-                merged_dict = self.sub(
-                    merged_dict,
-                    "NAM_DATA_ISBA",
-                    "CFNAM_H_TREE(@VEGTYPE@)",
-                    "ASCLLV",
-                    vtype=vtt,
+            fname = str(self.config.get_setting("SURFEX#COVER#YCOVER"))
+            __, filetype = self.get_filetype_from_suffix(fname)
+            if filetype is not None:
+                problems = self.check_nml_setting(
+                    problems, nml, "NAM_COVER", "YCOVERFILETYPE", filetype
                 )
 
-            key = "CFNAM_H_TREE(@VEGTYPE@)"
-            merged_dict = self.delete(merged_dict, "NAM_DATA_ISBA", key)
-            key = "CFTYP_H_TREE(@VEGTYPE@)"
-            merged_dict = self.delete(merged_dict, "NAM_DATA_ISBA", key)
-
-        return merged_dict
-
-    def set_prep_namelist(
-        self,
-        merged_dict,
-        prep_file=None,
-        prep_filetype=None,
-        prep_pgdfile=None,
-        prep_pgdfiletype=None,
-    ):
-        """Set prep namelist.
-
-        Args:
-            merged_dict(dict): Merged settings.
-            prep_file (_type_, optional): _description_. Defaults to None.
-            prep_filetype (_type_, optional): _description_. Defaults to None.
-            prep_pgdfile (_type_, optional): _description_. Defaults to None.
-            prep_pgdfiletype (_type_, optional): _description_. Defaults to None.
-
-        Raises:
-            FileNotFoundError: Main prep namelist input
-            RuntimeError: You must provide a DTG for prep
-            RuntimeError: Filetype for input to PREP is not set!
-            RuntimeError: Filetype for PGD input to PREP is not set
-
-        Returns:
-            merged_dict(dict): Merged settings.
-
-        """
-        if prep_file is not None and prep_filetype is None:
-            raise RuntimeError("Filetype for input to PREP is not set!")
-        if prep_pgdfile is not None and prep_pgdfiletype is None:
-            raise RuntimeError("Filetype for PGD input to PREP is not set!")
-
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/prep.json"
-        )
-        if prep_file is not None:
-            if prep_file.endswith(".json"):
-                if not os.path.exists(prep_file):
-                    raise FileNotFoundError(f"{prep_file} does not exist!")
-                merged_dict = self.merge_json_namelist_file(merged_dict, prep_file)
-            else:
-                fname = os.path.basename(prep_file)
-                merged_dict = self.sub(merged_dict, "NAM_PREP_SURF_ATM", "CFILE", fname)
-                merged_dict = self.sub(
-                    merged_dict, "NAM_PREP_SURF_ATM", "CFILETYPE", prep_filetype
-                )
-                if prep_pgdfile is not None:
-                    fname = os.path.basename(prep_pgdfile)
-                    merged_dict = self.sub(
-                        merged_dict, "NAM_PREP_SURF_ATM", "CFILEPGD", fname
-                    )
-                    merged_dict = self.sub(
-                        merged_dict, "NAM_PREP_SURF_ATM", "CFILEPGDTYPE", prep_pgdfiletype
-                    )
-                else:
-                    merged_dict = self.delete(
-                        merged_dict, "NAM_PREP_SURF_ATM", "CFILEPGD"
-                    )
-                    merged_dict = self.delete(
-                        merged_dict, "NAM_PREP_SURF_ATM", "CFILEPGDTYPE"
-                    )
-
-        if self.dtg is not None:
-            prep_time = self.dtg
-            merged_dict = self.sub(
-                merged_dict, "NAM_PREP_SURF_ATM", "NYEAR", int(prep_time.strftime("%Y"))
-            )
-            merged_dict = self.sub(
-                merged_dict, "NAM_PREP_SURF_ATM", "NMONTH", int(prep_time.strftime("%m"))
-            )
-            merged_dict = self.sub(
-                merged_dict, "NAM_PREP_SURF_ATM", "NDAY", int(prep_time.strftime("%d"))
-            )
-            merged_dict = self.sub(
-                merged_dict,
-                "NAM_PREP_SURF_ATM",
-                "XTIME",
-                float(prep_time.strftime("%H")) * 3600.0,
-            )
-        else:
-            raise RuntimeError("You must provide a DTG for prep")
-
-        if self.config.get_setting("SURFEX#TILES#SEA") == "SEAFLX":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/prep_seaflux.json"
-            )
-            sea_ice = self.config.get_setting("SURFEX#SEA#ICE")
-            merged_dict = self.sub(
-                merged_dict, "NAM_PREP_SEAFLUX", "CSEAICE_SCHEME", sea_ice
-            )
-            if sea_ice == "SICE":
-                merged_dict = self.merge_json_namelist_file(
-                    merged_dict, self.input_path + "/prep_sice.json"
+            # ZS
+            fname = str(self.config.get_setting("SURFEX#ZS#YZS"))
+            __, filetype = self.get_filetype_from_suffix(fname)
+            if filetype is not None:
+                problems = self.check_nml_setting(
+                    problems, nml, "NAM_ZS", "YZSFILETYPE", filetype
                 )
 
-        if self.config.get_setting("SURFEX#TILES#INLAND_WATER") == "FLAKE":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/prep_flake.json"
-            )
-            lclim_lake = self.config.get_setting("SURFEX#FLAKE#LCLIM")
-            merged_dict = self.sub(
-                merged_dict, "NAM_PREP_FLAKE", "LCLIM_LAKE", lclim_lake
-            )
-
-        if self.config.get_setting("SURFEX#TILES#NATURE") == "ISBA":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/prep_isba.json"
-            )
-            # Set extra ISBA-DIF properties (Not needed in prep?)
+            # Set ISBA properties
             if self.config.get_setting("SURFEX#ISBA#SCHEME") == "DIF":
-                merged_dict = self.merge_json_namelist_file(
-                    merged_dict, self.input_path + "/prep_isba_dif.json"
+                problems = self.check_nml_setting(
+                    problems, nml, "NAM_ISBA", "CISBA", "DIF"
+                )
+                problems = self.check_nml_setting(
+                    problems, nml, "NAM_ISBA", "NGROUND_LAYER", 14
+                )
+            elif self.config.get_setting("SURFEX#ISBA#SCHEME") == "3-L":
+                problems = self.check_nml_setting(
+                    problems, nml, "NAM_ISBA", "CISBA", "3-L"
+                )
+                problems = self.check_nml_setting(
+                    problems, nml, "NAM_ISBA", "NGROUND_LAYER", 3
+                )
+            elif self.config.get_setting("SURFEX#ISBA#SCHEME") == "2-L":
+                problems = self.check_nml_setting(
+                    problems, nml, "NAM_ISBA", "CISBA", "2-L"
+                )
+                problems = self.check_nml_setting(
+                    problems, nml, "NAM_ISBA", "NGROUND_LAYER", 2
                 )
 
-            # ISBA CANOPY
-            lisba_canopy = self.config.get_setting("SURFEX#ISBA#CANOPY")
-            merged_dict = self.sub(
-                merged_dict, "NAM_PREP_ISBA", "LISBA_CANOPY", lisba_canopy
+            # Set patches
+            problems = self.check_nml_setting(
+                problems,
+                nml,
+                "NAM_ISBA",
+                "NPATCH",
+                self.config.get_setting("SURFEX#ISBA#NPATCH"),
             )
 
-            # Snow
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/prep_isba_snow.json"
-            )
-            snow_scheme = self.config.get_setting("SURFEX#ISBA#SNOW")
-            merged_dict = self.sub(
-                merged_dict, "NAM_PREP_ISBA_SNOW", "CSNOW", snow_scheme
-            )
-            if self.config.get_setting("SURFEX#ISBA#SNOW") == "CRO":
-                merged_dict = self.merge_json_namelist_file(
-                    merged_dict, self.input_path + "/prep_isba_snow_crocus.json"
-                )
-
-        return merged_dict
-
-    def set_offline_namelist(self, merged_dict):
-        """Set offline namelist."""
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/offline.json"
-        )
-
-        if self.config.get_setting("SURFEX#IO#LSELECT"):
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/selected_output.json"
+            # Set MEB
+            problems = self.check_nml_setting(
+                problems,
+                nml,
+                "NAM_ISBA",
+                "LMEB",
+                self.config.get_setting("SURFEX#ISBA#MEB"),
             )
 
-        # SEAFLX settings
-        if self.config.get_setting("SURFEX#TILES#SEA") == "SEAFLX":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/offline_seaflux.json"
-            )
-            # Surface perturbations
-            pertflux = self.config.get_setting("SURFEX#SEA#PERTFLUX")
-            merged_dict = self.sub(merged_dict, "NAM_SEAFLUXn", "LPERTFLUX", pertflux)
-
-        if self.config.get_setting("SURFEX#TILES#INLAND_WATER") == "WATFLX":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/offline_watflux.json"
-            )
-
-        if self.config.get_setting("SURFEX#TILES#INLAND_WATER") == "FLAKE":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/offline_flake.json"
-            )
-
-        # ISBA settings
-        if self.config.get_setting("SURFEX#TILES#NATURE") == "ISBA":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/offline_isba.json"
-            )
-            pertsurf = self.config.get_setting("SURFEX#ISBA#PERTSURF")
-            merged_dict = self.sub(merged_dict, "NAM_ISBAn", "LPERTSURF", pertsurf)
-            xcgmax = self.config.get_setting("SURFEX#ISBA#XCGMAX", abort=False)
-            if xcgmax is not None:
-                merged_dict = self.sub(merged_dict, "NAM_ISBAn", "XCGMAX", xcgmax)
-            xcsmax = self.config.get_setting("SURFEX#ISBA#XCSMAX", abort=False)
-            if xcsmax is not None:
-                merged_dict = self.sub(merged_dict, "NAM_ISBAn", "XCSMAX", xcsmax)
-
-        # SSO
-        sso = self.config.get_setting("SURFEX#SSO#SCHEME")
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/offline_sso.json"
-        )
-        merged_dict = self.sub(merged_dict, "NAM_SSON", "CROUGH", sso)
-        if sso == "OROTUR":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/offline_sso_orotur.json"
-            )
-            merged_dict = self.sub(merged_dict, "NAM_ISBAn", "XSOROT", self.geo.xdx)
-
-        if self.config.get_setting("SURFEX#TILES#TOWN") == "TEB":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/offline_teb.json"
-            )
-
-        # Perturbed offline settings (are overridden when running binary)
-        # Make sure variables are existing and consistent
-        merged_dict = self.prepare_offline_perturbation(merged_dict)
-
-        # TODO the need for this in forecast must be removed!
-        merged_dict = self.set_obs(merged_dict)
-
-        # Climate setting
-        if self.config.get_setting("SURFEX#SEA#LVOLATILE_SIC"):
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/offline_seaice.json"
-            )
-            merged_dict = self.sub(merged_dict, "NAM_SEAICEn", "LVOLATILE_SIC", True)
-            merged_dict = self.sub(merged_dict, "NAM_SEAICEn", "XSIC_EFOLDING_TIME", 1.0)
-
-        return merged_dict
-
-    def prepare_offline_perturbation(self, merged_dict):
-        """Prepare for offline pertubations.
-
-        Args:
-            merged_dict (_type_): _description_
-
-        Returns:
-            _type_: _description_
-
-        """
-        if "NAM_VAR" not in merged_dict:
-            merged_dict.update({"NAM_VAR": {}})
-        if "NAM_IO_VARASSIM" not in merged_dict:
-            merged_dict.update({"NAM_IO_VARASSIM": {}})
-        merged_dict["NAM_VAR"].update({"NIVAR": 0})
-        merged_dict["NAM_IO_VARASSIM"].update({"LPRT": False})
-        if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "EKF":
-
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/offline_assim_pert.json"
-            )
-
-            merged_dict = self.sub(merged_dict, "NAM_ASSIM", "CASSIM_ISBA", "EKF")
-            nvar = 0
-            cvar_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#CVAR_M")
-            nncv = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#NNCV")
-            xtprt_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XTPRT_M")
-            xsigma_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XSIGMA_M")
-            for var, cvar_val in enumerate(cvar_m):
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "CVAR_M(@VAR@)", cvar_val, var=var + 1
-                )
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "NNCV(@VAR@)", nncv[var], var=var + 1
-                )
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "XTPRT_M(@VAR@)", xtprt_m[var], var=var + 1
-                )
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "XSIGMA_M(@VAR@)", xsigma_m[var], var=var + 1
-                )
-                if nncv[var] == 1:
-                    nvar += 1
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "XSIGMA_M(@VAR@)")
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "XTPRT_M(@VAR@)")
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "CVAR_M(@VAR@)")
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "NNCV(@VAR@)")
-            merged_dict = self.sub(merged_dict, "NAM_VAR", "NVAR", nvar)
-
-        if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "ENKF":
-
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/offline_assim_pert.json"
-            )
-            merged_dict = self.sub(merged_dict, "NAM_ASSIM", "CASSIM_ISBA", "ENKF")
-
-            nvar = 0
-            cvar_m = self.config.get_setting("SURFEX#ASSIM#ISBA#ENKF#CVAR_M")
-            nncv = self.config.get_setting("SURFEX#ASSIM#ISBA#ENKF#NNCV")
-            for var, cvar_val in enumerate(cvar_m):
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "CVAR_M(@VAR@)", cvar_val, var=var + 1
-                )
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "NNCV(@VAR@)", nncv[var], var=var + 1
-                )
-                if nncv[var] == 1:
-                    nvar += 1
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "XSIGMA_M(@VAR@)")
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "XTPRT_M(@VAR@)")
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "CVAR_M(@VAR@)")
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "NNCV(@VAR@)")
-            merged_dict = self.sub(merged_dict, "NAM_VAR", "NVAR", nvar)
-        return merged_dict
-
-    def set_obs(self, merged_dict):
-        """Set obs.
-
-        Args:
-            merged_dict (dict): Merged settings
-
-        Raises:
-            RuntimeError: Mismatch in nnco/cobs_m/xerrobs_m
-
-        Returns:
-            _type_: _description_
-
-        """
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/soda_obs.json"
-        )
-        lobsheader = self.config.get_setting("SURFEX#ASSIM#OBS#LOBSHEADER")
-        merged_dict = self.sub(merged_dict, "NAM_OBS", "LOBSHEADER", lobsheader)
-        lobsnat = self.config.get_setting("SURFEX#ASSIM#OBS#LOBSNAT")
-        merged_dict = self.sub(merged_dict, "NAM_OBS", "LOBSNAT", lobsnat)
-        cfile_format_obs = self.config.get_setting("SURFEX#ASSIM#OBS#CFILE_FORMAT_OBS")
-        merged_dict = self.sub(
-            merged_dict, "NAM_OBS", "CFILE_FORMAT_OBS", cfile_format_obs
-        )
-        nobstype = 0
-        nnco = self.config.get_setting("SURFEX#ASSIM#OBS#NNCO")
-        cobs_m = self.config.get_setting("SURFEX#ASSIM#OBS#COBS_M")
-        xerrobs_m = self.config.get_setting("SURFEX#ASSIM#OBS#XERROBS_M")
-        logging.debug("%s %s %s", nnco, cobs_m, xerrobs_m)
-        if len(nnco) != len(cobs_m) or len(nnco) != len(xerrobs_m):
-            raise RuntimeError("Mismatch in nnco/cobs_m/xerrobs_m")
-
-        for obs, obs_val in enumerate(nnco):
-            merged_dict = self.sub(
-                merged_dict, "NAM_OBS", "COBS_M(@VAR@)", cobs_m[obs], var=obs + 1
-            )
-            merged_dict = self.sub(
-                merged_dict, "NAM_OBS", "NNCO(@VAR@)", obs_val, var=obs + 1
-            )
-            merged_dict = self.sub(
-                merged_dict, "NAM_OBS", "XERROBS_M(@VAR@)", xerrobs_m[obs], var=obs + 1
-            )
-            if nnco[obs] == 1:
-                nobstype += 1
-        merged_dict = self.delete(merged_dict, "NAM_OBS", "COBS_M(@VAR@)")
-        merged_dict = self.delete(merged_dict, "NAM_OBS", "NNCO(@VAR@)")
-        merged_dict = self.delete(merged_dict, "NAM_OBS", "XERROBS_M(@VAR@)")
-        merged_dict = self.sub(merged_dict, "NAM_OBS", "NOBSTYPE", nobstype)
-        lswe = self.config.get_setting("SURFEX#ASSIM#OBS#LSWE")
-        merged_dict = self.sub(merged_dict, "NAM_OBS", "LSWE", lswe)
-        return merged_dict
-
-    def set_soda_namelist(self, merged_dict):
-        """Set SODA namelist.
-
-        Args:
-            merged_dict (dict): Merged dict
-
-        Raises:
-            RuntimeError: You must provide a DTG when using a list for snow
-            RuntimeError: Mismatch in nncv/cvar_m
-            RuntimeError: Mismatch in nncv/cvar_m/xsigma_m/xtprt_m
-
-        Returns:
-            dict: Merged dict.
-        """
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/soda.json"
-        )
-
-        merged_dict = self.sub(merged_dict, "NAM_ASSIM", "LASSIM", True)
-
-        # Set observations
-        merged_dict = self.set_obs(merged_dict)
-
-        # LSM
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_ASSIM",
-            "CFILE_FORMAT_LSM",
-            self.config.get_setting("SURFEX#ASSIM#CFILE_FORMAT_LSM"),
-        )
-
-        # Sea
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_ASSIM",
-            "CASSIM_SEA",
-            self.config.get_setting("SURFEX#ASSIM#SCHEMES#SEA"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_ASSIM",
-            "CFILE_FORMAT_SST",
-            self.config.get_setting("SURFEX#ASSIM#SEA#CFILE_FORMAT_SST"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_ASSIM",
-            "LREAD_SST_FROM_FILE",
-            self.config.get_setting("SURFEX#ASSIM#SEA#LREAD_SST_FROM_FILE"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_ASSIM",
-            "LEXTRAP_SEA",
-            self.config.get_setting("SURFEX#ASSIM#SEA#LEXTRAP_SEA"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_ASSIM",
-            "LECSST",
-            self.config.get_setting("SURFEX#ASSIM#SEA#LECSST"),
-        )
-
-        # Water
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_ASSIM",
-            "CASSIM_WATER",
-            self.config.get_setting("SURFEX#ASSIM#SCHEMES#INLAND_WATER"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_ASSIM",
-            "LWATERTG2",
-            self.config.get_setting("SURFEX#ASSIM#INLAND_WATER#LWATERTG2"),
-        )
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_ASSIM",
-            "LEXTRAP_WATER",
-            self.config.get_setting("SURFEX#ASSIM#INLAND_WATER#LEXTRAP_WATER"),
-        )
-
-        # Nature
-        merged_dict = self.sub(
-            merged_dict,
-            "NAM_ASSIM",
-            "CASSIM_ISBA",
-            self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA"),
-        )
-
-        # Snow
-        laesnm = False
-        snow_cycles = self.config.get_setting("SURFEX#ASSIM#ISBA#UPDATE_SNOW_CYCLES")
-        if len(snow_cycles) > 0:
-            logging.debug("%s", self.dtg)
-            if self.dtg is not None:
-                for cycle in snow_cycles:
-                    logging.debug(self.dtg.strftime("%H"))
-                    if int(self.dtg.strftime("%H")) == int(cycle):
-                        logging.debug("true")
-                        laesnm = True
-            else:
-                raise RuntimeError(
-                    "You must provide a DTG when using a list for snow  "
-                    "assimilation cycles"
-                )
-        merged_dict = self.sub(merged_dict, "NAM_ASSIM", "LAESNM", laesnm)
-
-        # Set OI settings
-        if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "OI":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/soda_isba_oi.json"
-            )
-            ua_physics = self.config.get_setting("FORECAST#PHYSICS", abort=False)
-            if ua_physics is None:
-                logging.warning("FORECAST#PHYSICS not set. Assume arome physics")
-                ua_physics = "arome"
-            if ua_physics == "arome":
-                merged_dict = self.sub(merged_dict, "NAM_ASSIM", "LAROME", True)
-            elif ua_physics == "alaro":
-                merged_dict = self.sub(merged_dict, "NAM_ASSIM", "LAROME", False)
-
-            merged_dict = self.sub(
-                merged_dict,
-                "NAM_NACVEG",
-                "XSIGT2MO",
-                self.config.get_setting("SURFEX#ASSIM#ISBA#OI#XSIGT2MO"),
-            )
-            merged_dict = self.sub(
-                merged_dict,
-                "NAM_NACVEG",
-                "XSIGH2MO",
-                self.config.get_setting("SURFEX#ASSIM#ISBA#OI#XSIGH2MO"),
-            )
-            merged_dict = self.sub(merged_dict, "NAM_NACVEG", "XRCLIMCA", 0.0)
-            merged_dict = self.sub(merged_dict, "NAM_NACVEG", "XRCLISST", 0.05)
-            merged_dict = self.sub(merged_dict, "NAM_NACVEG", "NECHGU", self.fcint)
-            merged_dict = self.sub(merged_dict, "NAM_NACVEG", "LOBS2M", True)
-            merged_dict = self.sub(merged_dict, "NAM_NACVEG", "LOBSWG", False)
-            f_clim = self.config.get_setting("SURFEX#ASSIM#ISBA#OI#CFILE_FORMAT_CLIM")
-            merged_dict = self.sub(merged_dict, "NAM_ASSIM", "CFILE_FORMAT_CLIM", f_clim)
-            f_fg = self.config.get_setting("SURFEX#ASSIM#ISBA#OI#CFILE_FORMAT_FG")
-            merged_dict = self.sub(merged_dict, "NAM_ASSIM", "CFILE_FORMAT_FG", f_fg)
-
-        if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "EKF":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/soda_isba_ekf.json"
-            )
-            nvar = 0
-            llincheck = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#LLINCHECK")
-            merged_dict = self.sub(merged_dict, "NAM_ASSIM", "LLINCHECK", llincheck)
-            xalpha = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XALPHA")
-            merged_dict = self.sub(merged_dict, "NAM_VAR", "XALPHA", xalpha)
-            cvar_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#CVAR_M")
-            xsigma_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XSIGMA_M")
-            xtprt_m = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XTPRT_M")
-            nncv = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#NNCV")
-            if (
-                len(nncv) != len(cvar_m)
-                or len(nncv) != len(xsigma_m)
-                or len(nncv) != len(xtprt_m)
-            ):
-                raise RuntimeError("Mismatch in nncv/cvar_m/xsigma_m/xtprt_m")
-            for var, cvar_val in enumerate(cvar_m):
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "CVAR_M(@VAR@)", cvar_val, var=var + 1
-                )
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "NNCV(@VAR@)", nncv[var], var=var + 1
-                )
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "XTPRT_M(@VAR@)", xtprt_m[var], var=var + 1
-                )
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "XSIGMA_M(@VAR@)", xsigma_m[var], var=var + 1
-                )
-                if nncv[var] == 1:
-                    nvar += 1
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "XSIGMA_M(@VAR@)")
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "XTPRT_M(@VAR@)")
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "CVAR_M(@VAR@)")
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "NNCV(@VAR@)")
-            merged_dict = self.sub(merged_dict, "NAM_VAR", "NIVAR", 0)
-            merged_dict = self.sub(merged_dict, "NAM_VAR", "NVAR", nvar)
-            xscale_q = self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#XSCALE_Q")
-            merged_dict = self.sub(merged_dict, "NAM_VAR", "XSCALE_Q", xscale_q)
-            merged_dict = self.sub(merged_dict, "NAM_IO_VARASSIM", "LPRT", False)
-            merged_dict = self.sub(
-                merged_dict,
-                "NAM_IO_VARASSIM",
-                "LBEV",
-                self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#EVOLVE_B"),
-            )
-            merged_dict = self.sub(
-                merged_dict,
-                "NAM_IO_VARASSIM",
-                "LBFIXED",
-                not self.config.get_setting("SURFEX#ASSIM#ISBA#EKF#EVOLVE_B"),
-            )
-
-        if self.config.get_setting("SURFEX#ASSIM#SCHEMES#ISBA") == "ENKF":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/soda_isba_enkf.json"
-            )
-            nvar = 0
-            cvar_m = self.config.get_setting("SURFEX#ASSIM#ISBA#ENKF#CVAR_M")
-            nncv = self.config.get_setting("SURFEX#ASSIM#ISBA#ENKF#NNCV")
-            if len(nncv) != len(cvar_m):
-                raise RuntimeError("Mismatch in nncv/cvar_m")
-            for var, cvar_val in enumerate(cvar_m):
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "CVAR_M(@VAR@)", cvar_val, var=var + 1
-                )
-                merged_dict = self.sub(
-                    merged_dict, "NAM_VAR", "NNCV(@VAR@)", nncv[var], var=var + 1
-                )
-
-                if nncv[var] == 1:
-                    nvar += 1
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "CVAR_M(@VAR@)")
-            merged_dict = self.delete(merged_dict, "NAM_VAR", "NNCV(@VAR@)")
-            merged_dict = self.sub(merged_dict, "NAM_VAR", "NIVAR", 0)
-            merged_dict = self.sub(merged_dict, "NAM_VAR", "NVAR", nvar)
-
-        # Town
-        town_setting = self.config.get_setting("SURFEX#ASSIM#SCHEMES#TEB")
-        merged_dict = self.sub(merged_dict, "NAM_ASSIM", "CASSIM_TEB", town_setting)
-        return merged_dict
-
-    def epilog(self, merged_dict):
-        """Epilog."""
-        # Always set these
-        if self.config.get_setting("SURFEX#SEA#ICE") == "SICE":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/epilog_sice.json"
-            )
-
-        merged_dict = self.merge_json_namelist_file(
-            merged_dict, self.input_path + "/epilog_treedrag.json"
-        )
-        lfaketrees = self.config.get_setting("SURFEX#TREEDRAG#FAKETREES", abort=False)
-        if lfaketrees is not None:
-            merged_dict = self.sub(merged_dict, "NAM_TREEDRAG", "LFAKETREE", lfaketrees)
-
-        if self.config.get_setting("SURFEX#TILES#INLAND_WATER") == "FLAKE":
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/epilog_flake.json"
-            )
-        return merged_dict
-
-    def override(self, merged_dict):
-        """Overide."""
-        # Override posssibility
-        if os.path.exists(self.input_path + "/override.json"):
-            logging.warning(
-                "WARNING: Override settings with content from %s/override.json",
-                self.input_path,
-            )
-            merged_dict = self.merge_json_namelist_file(
-                merged_dict, self.input_path + "/override.json"
-            )
-        return merged_dict
-
-    @staticmethod
-    def sub(merged_dict, nam_block, key, value, vtype=None, decade=None, var=None):
-        """Substitute key with value.
-
-        Args:
-            merged_dict (dict): Merged dict
-            nam_block (str): Namelist block
-            key (str): Key
-            value (any): Value
-            vtype(int, optional): Veg type. Defaults to None.
-            decade(int, optional): Decade type. Defaults to None.
-            var(int, optional): Var type. Defaults to None.
-
-        Returns:
-            dict: Merged dict
-
-        """
-        nam_block = nam_block.upper()
-        key = key.upper()
-        if nam_block in merged_dict:
-            if key in merged_dict[nam_block]:
-                setting = merged_dict[nam_block][key]
-                if vtype is not None:
-                    if isinstance(key, str):
-                        if "@VEGTYPE@" in setting:
-                            key = key.replace("@VEGTYPE@", str(vtype))
-                    if isinstance(setting, str):
-                        if "@VEGTYPE@" in setting:
-                            setting = setting.replace("@VEGTYPE@", str(vtype))
-                if decade is not None:
-                    if isinstance(key, str):
-                        if "@DECADE@" in setting:
-                            key = key.replace("@DECADE@", str(decade))
-                    if isinstance(setting, str):
-                        if "@DECADE@" in setting:
-                            setting = setting.replace("@DECADE@", str(decade))
-                if var is not None:
-                    if isinstance(key, str):
-                        if "@VAR@" in setting:
-                            key = key.replace("@VAR@", str(var))
-                    if isinstance(setting, str):
-                        if "@VAR@" in setting:
-                            setting = setting.replace("@VAR@", str(var))
-                logging.debug("key=%s, setting=%s", key, setting)
-                if isinstance(setting, str):
-                    if "@" + key + "@" in setting:
-                        setting = value
-                    else:
-                        logging.warning(
-                            "No string substitution. Using existing value: %s", setting
-                        )
-                merged_dict[nam_block][key] = setting
-            else:
-                logging.warning("No setting found in namelist for: %s", key)
+        elif self.program == "prep":
+            pass
+        elif self.program == "offline" or self.program == "perturbed":
+            pass
+        elif self.program == "soda":
+            pass
         else:
-            logging.warning("No namelist block found in namelist for: %s", nam_block)
-        return merged_dict
+            raise NotImplementedError(self.program)
 
-    @staticmethod
-    def delete(merged_dict, nam_block, key):
-        """Delete key from namelist block.
-
-        Args:
-            merged_dict (_type_): _description_
-            nam_block (_type_): _description_
-            key (_type_): _description_
-
-        Returns:
-            _type_: _description_
-
-        """
-        nam_block = nam_block.upper()
-        key = key.upper()
-        if nam_block in merged_dict:
-            if key in merged_dict[nam_block]:
-                del merged_dict[nam_block][key]
-        return merged_dict
+        return problems
 
     @staticmethod
     def get_filetype_from_suffix(fname):
@@ -2368,234 +433,142 @@ class Namelist(object):
             return fname, "ASCLLV"
         return fname, None
 
-    @staticmethod
-    def set_direct_data_namelist(lnamelist_section, ldtype, ldname, linput_path):
-        """Set direct data namelist.
-
-        Args:
-            lnamelist_section (_type_): _description_
-            ldtype (_type_): _description_
-            ldname (_type_): _description_
-            linput_path (_type_): _description_
-
-        Returns:
-            _type_: _description_
-
-        """
-        if ldname.endswith(".dir"):
-            basename = os.path.splitext(os.path.basename(ldname))[0]
-            filetype_name = ldtype
-            if ldtype == "YSOC_TOP" or ldtype == "YSOC_SUB":
-                filetype_name = "YSOC"
-            return {
-                "json": json.loads(
-                    '{"'
-                    + lnamelist_section
-                    + '": { "'
-                    + ldtype
-                    + '": "'
-                    + basename
-                    + '", '
-                    + '"'
-                    + filetype_name
-                    + 'FILETYPE": "DIRECT"}}'
-                )
-            }
-        if ldname.endswith(".json"):
-            return {"file": linput_path + "/" + ldname}
-
-    @staticmethod
-    def set_dirtyp_data_namelist(
-        lnamelist_section, ldtype, ldname, vtype=None, decade=None
+    def input_data_from_namelist(
+        self, input_data, platform, basetime=None, validtime=None
     ):
-        """Set dirtyp data namelist.
+        """Construct a base namelists class to be implemented by namelist implementations.
 
         Args:
-            lnamelist_section (_type_): _description_
-            ldtype (_type_): _description_
-            ldname (_type_): _description_
-            vtype (_type_, optional): _description_. Defaults to None.
-            decade (_type_, optional): _description_. Defaults to None.
+            input_data (dict): Input data definitions
+            platform (SystemFilePaths): Platform specific settings
+            basetime (as_datetime, optional): Base time
+            validtime (as_datetime, optional): Valid time
 
-        Returns:
-            _type_: _description_
+        Raises:
+            RuntimeError: Input
+            RuntimeError: Merged dictionary contains a @ in value
+            NotImplementedError: Mode is not implemented
 
         """
-        basename = os.path.splitext(os.path.basename(ldname))[0]
-        filetype_name = ldtype.upper()
-        if vtype is not None or decade is not None:
-            filetype_name = filetype_name + "("
-        if vtype is not None:
-            filetype_name = filetype_name + str(vtype)
-        if vtype is not None and decade is not None:
-            filetype_name = filetype_name + ","
-        if decade is not None:
-            filetype_name = filetype_name + str(decade)
-        if vtype is not None or decade is not None:
-            filetype_name = filetype_name + ")"
-        return {
-            "json": json.loads(
-                '{"'
-                + lnamelist_section
-                + '": { "CFNAM_'
-                + filetype_name
-                + '": "'
-                + basename
-                + '", "CFTYP_'
-                + filetype_name
-                + '": "DIRTYPE"}}'
+        logging.info("Set input data from namelist for program: %s", self.program)
+
+        data_obj = InputDataFromNamelist(
+            self.nml,
+            input_data,
+            self.program,
+            platform,
+            basetime=basetime,
+            validtime=validtime,
+        )
+        return data_obj
+
+    def assemble_namelist(self):
+        """Generate the namelists for 'target'.
+
+        Returns:
+            nlres (dict): Assembled namelist
+
+        """
+        # Read namelist file with all the categories
+
+        # Check target is valid
+        cndict = self.assemble
+        nldict = self.nldict
+
+        # Start with empty result dictionary
+        nlres = {}
+
+        # Assemble the target namelists based on the given category order
+        for item in self.flatten_list(cndict):
+            catg = item
+            # variable substitution removed at this level (may be resurrected)
+            # assemble namelists for this category
+            if catg in nldict:
+                for nl in nldict[catg]:
+                    if nl not in nlres:
+                        # create the result namelist dict
+                        nlres[nl] = {}
+                    if catg == "rm{" + nl + "}":
+                        # clear/remove the given namelist (but not used for now)
+                        nlres[nl].clear()
+                    else:
+                        for key in nldict[catg][nl]:
+                            val = nldict[catg][nl][key]
+                            finval = val
+                            # Replace ${var-def} with value from config, possibly macro-expanded
+                            # For now assumes only one subst. per line, could be generalized if needed
+                            if str(finval).find("$") >= 0:
+                                m = re.search(
+                                    r"^([^\$]*)\$\{([\w\.]+)\-?([^}]*)\}(.*)", str(val)
+                                )
+                                if m:
+                                    pre = m.group(1)
+                                    nam = m.group(2)
+                                    defval = m.group(3)
+                                    post = m.group(4)
+                                    logging.debug("macros=%s", self.macros)
+                                    logging.debug("look for nam=%s", nam)
+                                    try:
+                                        repval = self.macros[nam]
+                                    except KeyError:
+                                        repval = None
+                                    if repval is None:
+                                        if defval != "":
+                                            logging.debug(
+                                                "Using default value %s for '%s'",
+                                                defval,
+                                                nam,
+                                            )
+                                            repval = self.find_num(defval)
+                                        else:
+                                            logging.debug("No value found for: '%s'", nam)
+                                    else:
+                                        logging.debug(
+                                            "Replaced %s with: %s", nam, str(repval)
+                                        )
+                                    if isinstance(repval, str):
+                                        finval = str(pre) + str(repval) + str(post)
+                                    else:
+                                        finval = repval
+                                else:
+                                    raise KeyError(val)
+                            nlres[nl][key] = finval
+        return nlres
+
+    def write(self, output_file):
+        """Generate the namelists for 'target'.
+
+        Args:
+            output_file (str): where to write the result (OPTIONS.nam, fort.4 or EXSEG1.nam typically)
+
+        """
+        self.nml.uppercase = True
+        self.nml.true_repr = ".TRUE."
+        self.nml.false_repr = ".FALSE."
+        self.nml.write(output_file, force=True)
+        logging.debug("Wrote: %s", output_file)
+
+    @staticmethod
+    def flatten_list(li):
+        """Recursively flatten a list of lists (of lists)."""
+        if li == []:
+            return li
+        if isinstance(li[0], list):
+            return NamelistGenerator.flatten_list(li[0]) + NamelistGenerator.flatten_list(
+                li[1:]
             )
-        }
+        return li[:1] + NamelistGenerator.flatten_list(li[1:])
 
     @staticmethod
-    def capitalize_namelist_dict(dict_in):
-        """Capitalize namelist.
-
-        Args:
-            dict_in (_type_): _description_
-
-        Returns:
-            _type_: _description_
-
-        """
-        new_dict = {}
-        for key in dict_in:
-            upper_case2 = {}
-            for key2 in dict_in[key]:
-                upper_case2.update({key2.upper(): dict_in[key][key2]})
-            new_dict.update({key.upper(): upper_case2})
-        return new_dict
-
-    @staticmethod
-    def lower_case_namelist_dict(dict_in):
-        """Lower case namelist.
-
-        Args:
-            dict_in (dict): Namelist dictionary to lower case
-
-        Returns:
-            dict: Namelist in lower case
-
-        """
-        new_dict = {}
-        for key in dict_in:
-            lower_case_dict = {}
-            for key2 in dict_in[key]:
-                lower_case_dict.update({key2.lower(): dict_in[key][key2]})
-            new_dict.update({key.lower(): lower_case_dict})
-        return new_dict
-
-    @staticmethod
-    def merge_namelist_dicts(old_dict, new_dict):
-        """Merge namelist dicts.
-
-        Args:
-            old_dict (dict): Old dictionary
-            new_dict (dict): New dictionary
-
-        Returns:
-            dict: Merged dictionary
-
-        """
-        old_dict = Namelist.capitalize_namelist_dict(old_dict)
-        new_dict = Namelist.capitalize_namelist_dict(new_dict)
-        merged_dict = old_dict
-
-        for new_key, new_val in new_dict.items():
-            # Namelist block already exists
-            if new_key in merged_dict:
-                settings = merged_dict[new_key]
-                for new_key2 in new_val:
-                    settings.update({new_key2: new_dict[new_key][new_key2]})
-
-                merged_dict.update({new_key: settings})
-            # New namelist block
-            else:
-                merged_dict.update({new_key: new_val})
-
-        return merged_dict
-
-    @staticmethod
-    def ascii2nml(input_data):
-        """Convert dict to a namelist object.
-
-        Args:
-            input_data (dict): Namelist settings
-
-        Returns:
-            f90nml.Namelist: Namelist object.
-        """
-        output_data = f90nml.Namelist(input_data)
-        return output_data
-
-    @staticmethod
-    def ascii_file2nml(input_fname, input_fmt="json"):
-        """Convert a file wih namelist settings to a Namelist object.
-
-        Args:
-            input_fname (str): Filname
-            input_fmt (str, optional): File format. Defaults to "json".
-
-        Returns:
-            f90nml.Namelist: Namelist object.
-
-        """
-        if input_fmt == "json":
-            with open(input_fname, mode="r", encoding="utf-8") as input_file:
-                output_data = json.load(input_file)
-        elif input_fmt == "yaml":
-            with open(input_fname, mode="r", encoding="utf-8") as input_file:
-                output_data = yaml.safe_load(input_file)
-        output_data = f90nml.Namelist(output_data)
-        return output_data
-
-    @staticmethod
-    def nml2ascii(input_data, output_file, output_fmt="json", indent=2):
-        """Dump a namelist object as a dict in a json or yaml file.
-
-        Args:
-            input_data (f90nml.Namelist): Namelist object.
-            output_file (str): Filename
-            output_fmt (str, optional): File format. Defaults to "json".
-            indent (int, optional): Indentation. Defaults to 2.
-
-        """
-        if output_fmt == "json":
-            input_data = input_data.todict(complex_tuple=True)
-            json.dump(
-                input_data,
-                open(output_file, "w", encoding="utf-8"),
-                indent=indent,
-                separators=(",", ": "),
-            )
-        elif output_fmt == "yaml":
-            input_data = input_data.todict(complex_tuple=True)
-            yaml.dump(input_data, output_file, default_flow_style=False)
-
-    @staticmethod
-    def merge_json_namelist_file(old_dict, my_file):
-        """Merge json files with namelists.
-
-        Args:
-            old_dict (dict): Exististing settings
-            my_file (str): Filename with new settings
-
-        Returns:
-            dict: Merged settings.
-
-        """
-        logging.debug(my_file)
-        if os.path.exists(my_file):
-            with open(my_file, "r", encoding="utf-8") as file_handler:
-                new_dict = json.load(file_handler)
-        else:
-            logging.info("Possible namelist json file %s does not exist", my_file)
-            new_dict = {}
-
-        return Namelist.merge_namelist_dicts(old_dict, new_dict)
-
-    def get_namelist(self):
-        """Get namelist."""
-        logging.debug("Constructing namelist:")
-        merged_json_settings = self.namelist_dict
-        return self.ascii2nml(merged_json_settings)
+    def find_num(s):
+        """Purpose: un-quote numbers."""
+        try:
+            i = int(s)
+            return i
+        except ValueError:
+            pass
+        try:
+            f = float(s)
+            return f
+        except ValueError:
+            return s
