+++ tests/smoke/test_cli_run_binary.py
@@ -80,13 +80,56 @@ def test_run_pgd(
     get_rte_file,
     get_system,
     conf_proj_2x3_file,
+    get_nam_file,
+    tmp_path_factory,
+    input_binary_data_file,
+):
+    """Test run NC."""
+    # PGD
+
+    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/PGD_test.nc"
+    binary = "touch PGD.nc"
+
+    argv = [
+        "-w",
+        "",
+        "-c",
+        get_nc_config_file,
+        "--domain",
+        conf_proj_2x3_file,
+        "-s",
+        get_system,
+        "-n",
+        get_nam_file,
+        "-i",
+        input_binary_data_file,
+        "-r",
+        get_rte_file,
+        "-f",
+        "--tolerate_missing",
+        "-o",
+        output,
+        binary,
+    ]
+    with working_directory(tmp_path_factory.getbasetemp()):
+        pgd(argv=argv)
+
+
+@pytest.mark.usefixtures("_mockers")
+def test_run_pgd_old(
+    get_nc_config_file,
+    get_rte_file,
+    get_system,
+    conf_proj_2x3_file,
     get_nam_path,
     tmp_path_factory,
 ):
     """Test run NC."""
     # PGD
 
-    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/PGD_test.nc"
+    wrk = tmp_path_factory.getbasetemp() / "nam_old"
+    wrk.mkdir(exist_ok=True)
+    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/PGD_old_test.nc"
     binary = "touch PGD.nc"
 
     argv = [
@@ -108,12 +151,57 @@ def test_run_pgd(
         output,
         binary,
     ]
-    with working_directory(tmp_path_factory.getbasetemp()):
+    with working_directory(wrk):
         pgd(argv=argv)
 
 
 @pytest.mark.usefixtures("_mockers")
 def test_run_prep(
+    get_nc_config_file,
+    get_system,
+    get_rte_file,
+    get_nam_file,
+    conf_proj_2x3_file,
+    tmp_path_factory,
+    input_binary_data_file,
+):
+
+    # PREP
+
+    pgd = tmp_path_factory.getbasetemp() / "PGD_input.nc"
+    pgd.touch()
+    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/PREP_test.nc"
+    binary = "touch PREP.nc"
+
+    argv = [
+        "-w",
+        "",
+        "--domain",
+        conf_proj_2x3_file,
+        "--pgd",
+        pgd.as_posix(),
+        "-c",
+        get_nc_config_file,
+        "-s",
+        get_system,
+        "-n",
+        get_nam_file,
+        "-i",
+        input_binary_data_file,
+        "-r",
+        get_rte_file,
+        "-f",
+        "--tolerate_missing",
+        "-o",
+        output,
+        binary,
+    ]
+    with working_directory(tmp_path_factory.getbasetemp()):
+        prep(argv=argv)
+
+
+@pytest.mark.usefixtures("_mockers")
+def test_run_prep_old(
     get_nc_config_file,
     get_system,
     get_rte_file,
@@ -123,10 +211,11 @@ def test_run_prep(
 ):
 
     # PREP
-
-    pgd = tmp_path_factory.getbasetemp() / "PGD_input.nc"
+    wrk = tmp_path_factory.getbasetemp() / "nam_old"
+    wrk.mkdir(exist_ok=True)
+    pgd = wrk / "PGD_input.nc"
     pgd.touch()
-    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/PREP_test.nc"
+    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/PREP_old_test.nc"
     binary = "touch PREP.nc"
 
     argv = [
@@ -156,7 +245,8 @@ def test_run_prep(
         output,
         binary,
     ]
-    with working_directory(tmp_path_factory.getbasetemp()):
+
+    with working_directory(wrk):
         prep(argv=argv)
 
 
@@ -165,9 +255,10 @@ def test_run_offline(
     get_nc_config_file,
     get_rte_file,
     get_system,
-    get_nam_path,
+    get_nam_file,
     tmp_path_factory,
     conf_proj_2x3_file,
+    input_binary_data_file,
 ):
     # OFFLINE
 
@@ -176,6 +267,57 @@ def test_run_offline(
     prep = tmp_path_factory.getbasetemp() / "PREP_input.nc"
     prep.touch()
     output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/SURFOUT_test.nc"
+    binary = "touch SURFOUT.nc"
+
+    argv = [
+        "-w",
+        "",
+        "--domain",
+        conf_proj_2x3_file,
+        "--pgd",
+        pgd.as_posix(),
+        "--prep",
+        prep.as_posix(),
+        "-c",
+        get_nc_config_file,
+        "-s",
+        get_system,
+        "-n",
+        get_nam_file,
+        "-i",
+        input_binary_data_file,
+        "-r",
+        get_rte_file,
+        "-f",
+        "--tolerate_missing",
+        "-o",
+        output,
+        "--forc_zs",
+        "--forcing_dir",
+        "testdata",
+        binary,
+    ]
+    with working_directory(tmp_path_factory.getbasetemp()):
+        offline(argv=argv)
+
+
+@pytest.mark.usefixtures("_mockers")
+def test_run_offline_old(
+    get_nc_config_file,
+    get_rte_file,
+    get_system,
+    get_nam_path,
+    tmp_path_factory,
+    conf_proj_2x3_file,
+):
+    # OFFLINE
+    wrk = tmp_path_factory.getbasetemp() / "nam_old"
+    wrk.mkdir(exist_ok=True)
+    pgd = wrk / "PGD_input.nc"
+    pgd.touch()
+    prep = wrk / "PREP_input.nc"
+    prep.touch()
+    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/SURFOUT_old_test.nc"
     binary = "touch SURFOUT.nc"
 
     argv = [
@@ -204,7 +346,7 @@ def test_run_offline(
         "testdata",
         binary,
     ]
-    with working_directory(tmp_path_factory.getbasetemp()):
+    with working_directory(wrk):
         offline(argv=argv)
 
 
@@ -213,15 +355,69 @@ def test_run_perturbed(
     get_nc_config_file,
     get_rte_file,
     get_system,
-    get_nam_path,
+    get_nam_file,
     tmp_path_factory,
     conf_proj_2x3_file,
+    input_binary_data_file,
 ):
     # PERTURBED OFFLINE
 
     pgd = tmp_path_factory.getbasetemp() / "PGD_input.nc"
     pgd.touch()
     prep = tmp_path_factory.getbasetemp() / "PREP_input.nc"
+    prep.touch()
+    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/SURFOUT_1_test.nc"
+    binary = "touch SURFOUT.nc"
+
+    argv = [
+        "-w",
+        "",
+        "--domain",
+        conf_proj_2x3_file,
+        "--pgd",
+        pgd.as_posix(),
+        "--prep",
+        prep.as_posix(),
+        "-c",
+        get_nc_config_file,
+        "-s",
+        get_system,
+        "-n",
+        get_nam_file,
+        "-i",
+        input_binary_data_file,
+        "-r",
+        get_rte_file,
+        "--pert",
+        "1",
+        "-f",
+        "--tolerate_missing",
+        "-o",
+        output,
+        "--forc_zs",
+        "--forcing_dir",
+        "testdata",
+        binary,
+    ]
+    with working_directory(tmp_path_factory.getbasetemp()):
+        perturbed_offline(argv=argv)
+
+
+@pytest.mark.usefixtures("_mockers")
+def test_run_perturbed_old(
+    get_nc_config_file,
+    get_rte_file,
+    get_system,
+    get_nam_path,
+    tmp_path_factory,
+    conf_proj_2x3_file,
+):
+    # PERTURBED OFFLINE
+    wrk = tmp_path_factory.getbasetemp() / "nam_old"
+    wrk.mkdir(exist_ok=True)
+    pgd = wrk / "PGD_input.nc"
+    pgd.touch()
+    prep = wrk / "PREP_input.nc"
     prep.touch()
     output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/SURFOUT_1_test.nc"
     binary = "touch SURFOUT.nc"
@@ -254,7 +450,7 @@ def test_run_perturbed(
         "testdata",
         binary,
     ]
-    with working_directory(tmp_path_factory.getbasetemp()):
+    with working_directory(wrk):
         perturbed_offline(argv=argv)
 
 
@@ -263,9 +459,10 @@ def test_run_soda(
     get_nc_config_file,
     get_system,
     get_rte_file,
-    get_nam_path,
+    get_nam_file,
     conf_proj_2x3_file,
     tmp_path_factory,
+    input_binary_data_file,
 ):
     # SODA
 
@@ -274,6 +471,54 @@ def test_run_soda(
     prep = tmp_path_factory.getbasetemp() / "PREP_input.nc"
     prep.touch()
     output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/ANALYSIS_test.nc"
+    binary = "touch SURFOUT.nc"
+
+    argv = [
+        "-w",
+        "",
+        "--domain",
+        conf_proj_2x3_file,
+        "--pgd",
+        pgd.as_posix(),
+        "--prep",
+        prep.as_posix(),
+        "-c",
+        get_nc_config_file,
+        "-s",
+        get_system,
+        "-n",
+        get_nam_file,
+        "-i",
+        input_binary_data_file,
+        "-r",
+        get_rte_file,
+        "-f",
+        "--tolerate_missing",
+        "-o",
+        output,
+        binary,
+    ]
+    with working_directory(tmp_path_factory.getbasetemp()):
+        soda(argv)
+
+
+@pytest.mark.usefixtures("_mockers")
+def test_run_soda_old(
+    get_nc_config_file,
+    get_system,
+    get_rte_file,
+    get_nam_path,
+    conf_proj_2x3_file,
+    tmp_path_factory,
+):
+    # SODA
+    wrk = tmp_path_factory.getbasetemp() / "nam_old"
+    wrk.mkdir(exist_ok=True)
+    pgd = wrk / "PGD_input.nc"
+    pgd.touch()
+    prep = wrk / "PREP_input.nc"
+    prep.touch()
+    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/ANALYSIS_old_test.nc"
     binary = "touch SURFOUT.nc"
 
     argv = [
@@ -301,7 +546,7 @@ def test_run_soda(
         output,
         binary,
     ]
-    with working_directory(tmp_path_factory.getbasetemp()):
+    with working_directory(wrk):
         soda(argv)
 
 
@@ -310,9 +555,10 @@ def test_masterodb_forecast(
     get_fa_config_file,
     get_system,
     get_rte_file,
-    get_nam_path,
+    get_nam_file,
     conf_proj_2x3_file,
     tmp_path_factory,
+    input_binary_data_file,
 ):
     """Test masterodb."""
     pgd = tmp_path_factory.getbasetemp() / "Const.Clim.sfx"
@@ -321,6 +567,57 @@ def test_masterodb_forecast(
     prep.touch()
     output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/ICMSHHARM+0003.sfx"
     binary = "touch ICMSHHARM+0003.fa"
+
+    argv = [
+        "-w",
+        "",
+        "-m",
+        "forecast",
+        "--domain",
+        conf_proj_2x3_file,
+        "--pgd",
+        pgd.as_posix(),
+        "--prep",
+        prep.as_posix(),
+        "-c",
+        get_fa_config_file,
+        "-s",
+        get_system,
+        "-n",
+        get_nam_file,
+        "-i",
+        input_binary_data_file,
+        "-r",
+        get_rte_file,
+        "-f",
+        "--tolerate_missing",
+        "-o",
+        output,
+        "-b",
+        binary,
+    ]
+    with working_directory(tmp_path_factory.getbasetemp()):
+        masterodb(argv=argv)
+
+
+@pytest.mark.usefixtures("_mockers")
+def test_masterodb_forecast_old(
+    get_fa_config_file,
+    get_system,
+    get_rte_file,
+    get_nam_path,
+    conf_proj_2x3_file,
+    tmp_path_factory,
+):
+    """Test masterodb."""
+    wrk = tmp_path_factory.getbasetemp() / "nam_old"
+    wrk.mkdir(exist_ok=True)
+    pgd = wrk / "Const.Clim.sfx"
+    pgd.touch()
+    prep = wrk / "ICMSHHARMINIT.sfx"
+    prep.touch()
+    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/ICMSHHARM+0003_old.sfx"
+    binary = "touch ICMSHHARM+0003_old.fa"
 
     argv = [
         "-w",
@@ -348,7 +645,7 @@ def test_masterodb_forecast(
         "-b",
         binary,
     ]
-    with working_directory(tmp_path_factory.getbasetemp()):
+    with working_directory(wrk):
         masterodb(argv=argv)
 
 
@@ -357,17 +654,68 @@ def test_masterodb_canari(
     get_fa_config_file,
     get_system,
     get_rte_file,
+    get_nam_file,
+    conf_proj_2x3_file,
+    tmp_path_factory,
+    input_binary_data_file,
+):
+    # CANARI
+    pgd = tmp_path_factory.getbasetemp() / "Const.Clim.sfx"
+    pgd.touch(exist_ok=True)
+    prep = tmp_path_factory.getbasetemp() / "ICMSHHARMINIT.sfx"
+    prep.touch()
+    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/ICMSHANAL+0000.sfx"
+    binary = "touch ICMSHANAL.sfx"
+    argv = [
+        "-w",
+        "",
+        "-m",
+        "canari",
+        "--domain",
+        conf_proj_2x3_file,
+        "--pgd",
+        pgd.as_posix(),
+        "--prep",
+        prep.as_posix(),
+        "-c",
+        get_fa_config_file,
+        "-s",
+        get_system,
+        "-n",
+        get_nam_file,
+        "-i",
+        input_binary_data_file,
+        "-r",
+        get_rte_file,
+        "-f",
+        "--tolerate_missing",
+        "-o",
+        output,
+        "-b",
+        binary,
+    ]
+    with working_directory(tmp_path_factory.getbasetemp()):
+        masterodb(argv=argv)
+
+
+@pytest.mark.usefixtures("_mockers")
+def test_masterodb_canari_old(
+    get_fa_config_file,
+    get_system,
+    get_rte_file,
     get_nam_path,
     conf_proj_2x3_file,
     tmp_path_factory,
 ):
     # CANARI
-    pgd = tmp_path_factory.getbasetemp() / "Const.Clim.sfx"
+    wrk = tmp_path_factory.getbasetemp() / "nam_old"
+    wrk.mkdir(exist_ok=True)
+    pgd = wrk / "Const.Clim.sfx"
     pgd.touch()
-    prep = tmp_path_factory.getbasetemp() / "ICMSHHARMINIT.sfx"
+    prep = wrk / "ICMSHHARMINIT.sfx"
     prep.touch()
-    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/ICMSHHARM+0003.sfx"
-    binary = "touch ICMSHANAL.sfx"
+    output = f"{tmp_path_factory.getbasetemp().as_posix()}/archive/ICMSHANAL+0000_old.sfx"
+    binary = "touch ICMSHANAL_old.sfx"
     argv = [
         "-w",
         "",
@@ -396,5 +744,5 @@ def test_masterodb_canari(
         "-b",
         binary,
     ]
-    with working_directory(tmp_path_factory.getbasetemp()):
+    with working_directory(wrk):
         masterodb(argv=argv)
